3D-Dimensional reduction plot for cell classification visualization.

Plotting cell points on a reduced 3D space and coloring according to the groups of the cells.

Usage

ClassDimPlot3D(
  srt,
  group.by = "orig.ident",
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  palette = "Paired",
  pal_color = NULL,
  bg_color = "grey80",
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  lineages = NULL,
  lineage_palette = "Dark2",
  span = 0.75,
  arrow_reverse = FALSE,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)

Arguments

srt

A Seurat object.

group.by

Name of one or more metadata columns to group (color) cells by (for example, orig.ident).

reduction

Which dimensionality reduction to use.

palette

Name of palette to use. Default is "Paired".

bg_color

Color value for NA points.

pt.size

Point size for plotting.

cells.highlight

A vector of names of cells to highlight.

cols.highlight

A vector of colors to highlight the cells.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

sizes.highlight

Size of highlighted cells.

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

force

Examples

data(pancreas1k)
pancreas1k <- Standard_SCP(pancreas1k)
#> [2022-06-29 04:15:40] Start Standard_SCP
#> [2022-06-29 04:15:40] Checking srtList... ...
#> Data 1/1 of the srtList is raw counts. Perform NormalizeData(logCPM) on the data ...
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from the existed HVF in srtList...
#> [2022-06-29 04:15:43] Finished checking.
#> Perform ScaleData on the data...
#> Perform linear dimension reduction (pca) on the data...
#> Perform FindClusters (louvain) on the data...
#> Reorder clusters...
#> Perform nonlinear dimension reduction (umap) on the data...
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> [2022-06-29 04:16:00] Standard_SCP done
#> Elapsed time: 19.63 secs 
ClassDimPlot3D(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")

pancreas1k <- Standard_SCP(pancreas1k)
#> [2022-06-29 04:16:00] Start Standard_SCP
#> [2022-06-29 04:16:00] Checking srtList... ...
#> Data 1/1 of the srtList has been log-normalized.
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from the existed HVF in srtList...
#> [2022-06-29 04:16:02] Finished checking.
#> Perform ScaleData on the data...
#> Perform linear dimension reduction (pca) on the data...
#> Perform FindClusters (louvain) on the data...
#> Reorder clusters...
#> Perform nonlinear dimension reduction (umap) on the data...
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> [2022-06-29 04:16:19] Standard_SCP done
#> Elapsed time: 19.19 secs 
pancreas1k <- RunSlingshot(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
ClassDimPlot3D(pancreas1k, group.by = "SubCellType", reduction = "StandardpcaUMAP3D", lineages = paste0("Lineage", 1:2))