CreateMetaFile

Creates a meta file in HDF5 format from a Seurat object.

Usage

CreateMetaFile(
  srt,
  MetaFile,
  name = NULL,
  write_tools = FALSE,
  write_misc = FALSE,
  ignore_nlevel = 100,
  compression_level = 6,
  overwrite = FALSE
)

Arguments

srt

The Seurat object.

MetaFile

Path to the output meta file. If not provided, the file will be named "Meta.hdf5" in the current directory.

name

Name of the dataset. If not provided, the name will default to the Seurat object's project name.

write_tools

A logical value indicating whether to write the tools information to the meta file. Default is FALSE.

write_misc

A logical value indicating whether to write the miscellaneous information to the meta file. Default is FALSE.

ignore_nlevel

The number of levels above which a metadata field will be ignored. Default is 100.

compression_level

The level of compression for the meta file. Default is 6.

overwrite

A logical value indicating whether to overwrite existing metadata and reductions in the meta file. Default is FALSE.

@seealso CreateDataFile PrepareSCExplorer FetchH5 RunSCExplorer