Env_requirements function

This function provides the SCP python environment requirements for a specific version.

Usage

Env_requirements(version = "3.8-1")

Arguments

version

A character vector specifying the version of the environment (default is "3.8-1").

Value

A list of requirements for the specified version.

Details

The function returns a list of requirements including the required Python version and a list of packages with their corresponding versions.

Examples

# Get requirements for version "3.8-1"
Env_requirements("3.8-1")
#> $python
#> [1] "3.8"
#> 
#> $packages
#>               leidenalg              matplotlib                   numba                   numpy                palantir 
#>     "leidenalg==0.10.1"     "matplotlib==3.6.3"         "numba==0.55.2"         "numpy==1.21.6"       "palantir==1.0.1" 
#>                  pandas           python-igraph                  scanpy            scikit-learn                   scipy 
#>         "pandas==1.3.5" "python-igraph==0.10.2"         "scanpy==1.9.5"   "scikit-learn==1.3.2"         "scipy==1.10.1" 
#>                  scvelo                     wot                  trimap                  pacmap                   phate 
#>         "scvelo==0.2.5"      "wot==1.0.8.post2"         "trimap==1.1.4"         "pacmap==0.7.0"         "phate==1.0.11" 
#>                   bbknn               scanorama              scvi-tools 
#>          "bbknn==1.6.0"      "scanorama==1.7.4"    "scvi-tools==0.20.3" 
#>