Gene ID conversion function using biomart

This function can convert different gene ID types within one species or bewteen two species using biomart service.

Usage

GeneConvert(
  geneID,
  geneID_from_IDtype = "symbol",
  geneID_to_IDtype = "entrez_id",
  species_from = "Homo_sapiens",
  species_to = NULL,
  Ensembl_version = 103,
  try_times = 5,
  mirror = NULL
)

Arguments

geneID

A vector of the geneID character.

geneID_from_IDtype

ID type of the geneID. e.g. "symbol", "ensembl_id", "entrez_id"

geneID_to_IDtype

ID type to convert to. e.g. "symbol", "ensembl_id", "entrez_id"

species_from

Latin names for animals of the input geneID.

species_to

Latin names for animals of the output geneID. e.g. "Homo_sapiens","Mus_musculus"

Ensembl_version

Ensembl database version. If NULL, use the current release version.

try_times

Times when trying to connect with the biomart service.

mirror

Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.

Value

A list.

Examples

# \donttest{
res <- GeneConvert(
  geneID = c("CDK1", "MKI67", "TOP2A", "AURKA", "CTCF"), geneID_from_IDtype = "symbol", geneID_to_IDtype = "entrez_id",
  species_from = "Homo_sapiens", species_to = "Mus_musculus", Ensembl_version = 103
)
#> Connecting to the Ensembl archives...
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [www.ensembl.org] Connection timed out after 10001 milliseconds
str(res)
#> Error in str(res): object 'res' not found
# }