Prepare the gene annotation databases

Prepare the gene annotation databases

Usage

PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"),
  db = c("GO", "GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome",
    "ProteinComplex", "DGI", "MP", "DO", "PFAM", "Chromosome", "GeneType", "Enzyme",
    "TF", "SP", "CellTalk", "CellChat"),
  db_IDtypes = c("symbol", "entrez_id", "ensembl_id"),
  db_version = "latest",
  db_update = FALSE,
  convert_species = TRUE,
  Ensembl_version = 103,
  mirror = NULL,
  biomart = NULL,
  attempts = 5,
  custom_TERM2GENE = NULL,
  custom_TERM2NAME = NULL,
  custom_species = NULL,
  custom_IDtype = NULL,
  custom_version = NULL
)

Arguments

species

species

db

Enrichment database name.

db_update

Whether update the database to the latest version.

custom_version

Value

A list containing the database.

Examples

if (interactive()) {
  db_list <- PrepareDB(species = "Homo_sapiens", db = "GO_BP")
  ListDB(species = "Homo_sapiens", db = "GO_BP")
  head(db_list[["Homo_sapiens"]][["GO_BP"]][["TERM2GENE"]])

  # Based on homologous gene conversion, prepare a gene annotation database that originally does not exist in the species.
  db_list <- PrepareDB(species = "Homo_sapiens", db = "MP", convert_species = TRUE)
  ListDB(species = "Homo_sapiens", db = "MP")
  head(db_list[["Homo_sapiens"]][["MP"]][["TERM2GENE"]])

  # Prepare databases for other species
  db_list <- PrepareDB(species = "Macaca_fascicularis", db = "GO_BP", convert_species = TRUE)
  ListDB(species = "Macaca_fascicularis", db = "GO_BP")
  head(db_list[["Macaca_fascicularis"]][["GO_BP"]][["TERM2GENE"]])

  # Prepare databases for Arabidopsis (plant)
  db_list <- PrepareDB(
    species = "Arabidopsis_thaliana",
    db = c(
      "GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway",
      "ENZYME", "Chromosome"
    ),
    biomart = "plants_mart"
  )
  head(db_list[["Arabidopsis_thaliana"]][["KEGG"]][["TERM2GENE"]])

  # You can also build a custom database based on the gene sets you have
  ccgenes <- CC_GenePrefetch("Homo_sapiens")
  custom_TERM2GENE <- rbind(
    data.frame(term = "S_genes", gene = ccgenes[["cc_S_genes"]]),
    data.frame(term = "G2M_genes", gene = ccgenes[["cc_G2M_genes"]])
  )
  str(custom_TERM2GENE)

  # Set convert_species = TRUE to build a custom database for both species, with the name "CellCycle"
  db_list <- PrepareDB(
    species = c("Homo_sapiens", "Mus_musculus"), db = "CellCycle", convert_species = TRUE,
    custom_TERM2GENE = custom_TERM2GENE, custom_species = "Homo_sapiens", custom_IDtype = "symbol", custom_version = "Seurat_v4"
  )
  ListDB(db = "CellCycle")

  db_list <- PrepareDB(species = "Mus_musculus", db = "CellCycle")
  head(db_list[["Mus_musculus"]][["CellCycle"]][["TERM2GENE"]])
}