Run multi-dimensional scaling(MDS)

Run a MDS dimensionality reduction.

RunMDS(object, ...)

# S3 method for default
RunMDS(
  object,
  assay = NULL,
  slot = "scale.data",
  nmds = 50,
  dist.method = "cosine",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "MDS_",
  seed.use = 42,
  verbose = TRUE,
  ...
)

# S3 method for Assay
RunMDS(
  object,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  nmds = 50,
  dist.method = "cosine",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "MDS_",
  seed.use = 42,
  ...
)

# S3 method for Seurat
RunMDS(
  object,
  assay = NULL,
  slot = "scale.data",
  features = NULL,
  nmds = 50,
  dist.method = "cosine",
  mds.method = "cmdscale",
  rev.mds = FALSE,
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "mds",
  reduction.key = "MDS_",
  seed.use = 42,
  ...
)

Arguments

object

An object

assay

Name of Assay mds is being run on.

slot

Name of slot mds is being run on.

nmds

Total Number of MDS to compute and store (50 by default).

rev.mds

By default computes the mds on the cell x gene matrix. Setting to true will compute it on gene x cell matrix.

ndims.print

MDS to print genes for.

nfeatures.print

Number of genes to print for each MDS.

reduction.key

dimensional reduction key, specifies the string before the number for the dimension names. "MDS_" by default.

seed.use

Set a random seed. By default, sets the seed to 42. Setting NULL will not set a seed.

verbose

Print the top genes associated with high/low loadings for the MDS.

features

Features to compute mds on. If features=NULL, mds will be run using the variable features for the Assay.

reduction.name

dimensional reduction name, 'mds' by default

Value

Returns Seurat object with the MDS calculation stored in the reductions slot.