Run Palantir analysis

Run Palantir analysis

RunPalantir(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  h5ad = NULL,
  group_by = NULL,
  liner_reduction = NULL,
  nonliner_reduction = NULL,
  basis = NULL,
  n_pcs = 30,
  n_neighbors = 30,
  dm_n_components = 10,
  dm_alpha = 0,
  dm_n_eigs = NULL,
  early_group = NULL,
  terminal_groups = NULL,
  early_cell = NULL,
  terminal_cells = NULL,
  num_waypoints = 1200,
  scale_components = TRUE,
  use_early_cell_as_start = FALSE,
  max_iterations = 25,
  n_jobs = 8,
  point_size = 20,
  show_plot = TRUE,
  dpi = 300,
  save = FALSE,
  dirpath = "./",
  fileprefix = "",
  return_seurat = FALSE
)

Arguments

srt

A Seurat object.

adata

An anndata object. Generally created through srt_to_adata

h5ad

h5ad file path.

group_by

group_by.

liner_reduction

liner_reduction.

nonliner_reduction

nonliner_reduction.

basis

basis.

n_pcs

n_pcs.

n_neighbors

n_neighbors.

n_jobs

n_jobs.

point_size

point_size.

dpi

dpi.

save

save.

dirpath

dirpath.

fileprefix

fileprefix.

Value

A anndata object.

Examples

data("pancreas1k")
pancreas1k <- RunPalantir(
  srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP",
  early_group = "Ductal", terminal_groups = c("Alpha", "Beta", "Delta", "Epsilon"), return_seurat = TRUE
)
#> 'misc' slot in the srt object is not converted.
#> 'tools' slot in the srt object is not converted.
#> Warning: Keys should be one or more alphanumeric characters followed by an underscore, setting key from palantir_dm_ to palantirdm_
#> Warning: All keys should be one or more alphanumeric characters followed by an underscore '_', setting key to palantirdm_