Run scVelo analysis

scVelo is a scalable toolkit for RNA velocity analysis in single cells. This function performs scVelo workflow in R by reticulate.

RunSCVELO(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  h5ad = NULL,
  group_by = NULL,
  n_jobs = 1,
  liner_reduction = NULL,
  nonliner_reduction = NULL,
  basis = NULL,
  mode = "stochastic",
  fitting_by = "stochastic",
  magic_impute = FALSE,
  knn = 5,
  t = 2,
  min_shared_counts = 30,
  n_pcs = 30,
  n_neighbors = 30,
  approx = TRUE,
  stream_smooth = NULL,
  stream_density = 2,
  arrow_length = 5,
  arrow_size = 5,
  arrow_density = 0.5,
  denoise = FALSE,
  denoise_topn = 3,
  kinetics = FALSE,
  kinetics_topn = 100,
  calculate_velocity_genes = FALSE,
  s_genes = NULL,
  g2m_genes = NULL,
  show_plot = TRUE,
  dpi = 300,
  save = FALSE,
  dirpath = "./",
  fileprefix = "",
  return_seurat = FALSE
)

Arguments

srt

A Seurat object.

adata

An anndata object. Generally created through srt_to_adata

h5ad

h5ad file path.

group_by

group_by.

n_jobs

n_jobs.

liner_reduction

liner_reduction.

nonliner_reduction

nonliner_reduction.

basis

basis.

mode

mode.

fitting_by

fitting_by.

min_shared_counts

min_shared_counts.

n_pcs

n_pcs.

n_neighbors

n_neighbors.

approx

approx.

stream_smooth

stream_smooth.

stream_density

stream_density.

arrow_length

arrow_length.

arrow_size

arrow_size.

arrow_density

arrow_density.

calculate_velocity_genes

calculate_velocity_genes.

s_genes

s_genes.

g2m_genes

g2m_genes.

dpi

dpi.

save

save.

dirpath

dirpath.

fileprefix

fileprefix.

velocity_with_noise

velocity_with_noise.

diff_kinetics

diff_kinetics.

Value

A anndata object.

Examples

data("pancreas1k")
pancreas1k <- RunSCVELO(srt = pancreas1k, group_by = "SubCellType", liner_reduction = "PCA", nonliner_reduction = "UMAP", return_seurat = TRUE)
#> 'misc' slot in the srt object is not converted.
#> 'tools' slot in the srt object is not converted.
#> Warning: Keys should be one or more alphanumeric characters followed by an underscore, setting key from variance_stochastic_ to variancestochastic_
#> Warning: Keys should be one or more alphanumeric characters followed by an underscore, setting key from variance_velocity_ to variancevelocity_
#> Warning: Keys should be one or more alphanumeric characters followed by an underscore, setting key from X_UMAP_ to XUMAP_
#> Warning: All keys should be one or more alphanumeric characters followed by an underscore '_', setting key to XUMAP_
#> Warning: Keys should be one or more alphanumeric characters followed by an underscore, setting key from stochastic_UMAP_ to stochasticUMAP_
#> Warning: All keys should be one or more alphanumeric characters followed by an underscore '_', setting key to stochasticUMAP_
#> Warning: 'uns: neighbors' will not be converted. It is a Python object connected with the Python session. You may need to convert manually.