Convert an anndata object to a seurat object using reticulate

Convert an anndata object to a seurat object using reticulate

Usage

adata_to_srt(adata)

Arguments

adata

a connected python anndata object.

Examples

data("pancreas_sub")
adata <- srt_to_adata(pancreas_sub)
#> Error in check_Python(c("scanpy", "numpy")): could not find function "check_Python"
adata <- RunPAGA(adata = adata, group_by = "SubCellType", linear_reduction = "PCA", nonlinear_reduction = "UMAP")
#> Error in check_Python("scanpy"): could not find function "check_Python"
srt <- adata_to_srt(adata)
#> Error: object 'adata' not found
srt
#> Error: object 'srt' not found

### Or convert a h5ad file to Seurat object
# library(reticulate)
# check_Python("scanpy")
# sc <- import("scanpy")
# adata <- sc$read_h5ad("pancreas.h5ad")
# srt <- adata_to_srt(adata)
# srt