Check and install R packages

Check and install R packages

Usage

check_R(
  packages,
  package_names = NULL,
  install_methods = c("BiocManager::install", "install.packages",
    "devtools::install_github"),
  lib = .libPaths()[1],
  force = FALSE
)

Arguments

packages

Package to be installed. Package source can be CRAN, Bioconductor or Github, e.g. scmap, davidsjoberg/ggsankey.

package_names

The name of the package that corresponds to the packages parameter, used to check if the package is already installed. By default, the package name is extracted according to the packages parameter.

install_methods

Functions used to install R packages.

lib

The location of the library directories where to install the packages.

force

Whether to force the installation of packages. Default is FALSE.