Compute FDR from posterior error probabilities PEP

The functions takes the posterior error probabilities (PEPs) from the given assay's rowData and adds a new variable to it (called .FDR) that contains the computed false discovery rates (FDRs).

computeFDR(object, i, groupCol, pepCol)

Arguments

object
A QFeatures object

i
A numeric() or character() vector indicating from which assays the rowData should be taken.

groupCol
A character(1) indicating the variable names in the rowData that contains the grouping variable. The FDR are usually computed for PSMs grouped by peptide ID.

pepCol
A character(1) indicating the variable names in the rowData that contains the PEPs. Since, PEPs are probabilities, the variable must be contained in (0, 1).

Value

A QFeatures object.

Examples

data("scp1")
scp1 <- computeFDR(scp1,
                   i = 1,
                   groupCol = "Sequence",
                   pepCol = "dart_PEP")
## Check results
rowDataToDF(scp1, 1, c("dart_PEP", ".FDR"))

#> DataFrame with 195 rows and 4 columns
#>              dart_PEP        .FDR        .assay    .rowname
#>             <numeric>   <numeric>   <character> <character>
#> PSM116299 2.51576e-03 2.51576e-03 190321S_LC...   PSM116299
#> PSM116943 1.00000e+00 1.00000e+00 190321S_LC...   PSM116943
#> PSM117524 1.24305e-11 1.24305e-11 190321S_LC...   PSM117524
#> PSM119466 1.79900e-01 1.79900e-01 190321S_LC...   PSM119466
#> PSM119691 8.39315e-05 8.39315e-05 190321S_LC...   PSM119691
#> ...               ...         ...           ...         ...
#> PSM253842 9.00309e-04 9.00309e-04 190321S_LC...   PSM253842
#> PSM253848 5.55493e-06 5.55493e-06 190321S_LC...   PSM253848
#> PSM255286 2.53061e-02 2.53061e-02 190321S_LC...   PSM255286
#> PSM255919 5.45833e-12 5.45833e-12 190321S_LC...   PSM255919
#> PSM256061 2.21415e-05 2.21415e-05 190321S_LC...   PSM256061