Compute the sample over carrier ratio (SCR)

The function computes the ratio of the intensities of sample channels over the intentisty of the carrier channel for each feature. The ratios are averaged within the assay.

computeSCR(object, i, colDataCol, samplePattern, carrierPattern)

Arguments

object
A QFeatures object.

i
A character() or integer() indicating for which assay(s) the SCR needs to be computed.

colDataCol
A character(1) indicating the variable to take from colData(object) that gives the sample annotation.

samplePattern
A character(1) pattern that matches the sample encoding in colDataCol.

carrierPattern
A character(1) pattern that matches the carrier encoding in colDataCol. Only one match per assay is allowed, otherwise only the first match is taken

Value

A QFeatures object for which the rowData of the given assay(s) is augmented with the mean SCR (.meanSCR variable).

Examples

data("scp1")
scp1 <- computeSCR(scp1, 
                   i = 1,
                   colDataCol = "SampleType",
                   carrierPattern = "Carrier",
                   samplePattern = "Blank|Macrophage|Monocyte")
## Check results
rowDataToDF(scp1, 1, ".meanSCR")

#> DataFrame with 195 rows and 3 columns
#>             .meanSCR        .assay    .rowname
#>            <numeric>   <character> <character>
#> PSM116299  0.0868734 190321S_LC...   PSM116299
#> PSM116943  0.0328429 190321S_LC...   PSM116943
#> PSM117524  0.0160611 190321S_LC...   PSM117524
#> PSM119466        NaN 190321S_LC...   PSM119466
#> PSM119691  0.0118370 190321S_LC...   PSM119691
#> ...              ...           ...         ...
#> PSM253842 0.01138264 190321S_LC...   PSM253842
#> PSM253848 0.00966447 190321S_LC...   PSM253848
#> PSM255286 1.02935211 190321S_LC...   PSM255286
#> PSM255919 0.01340392 190321S_LC...   PSM255919
#> PSM256061 0.01092895 190321S_LC...   PSM256061