Compute the pairwise Jaccard index

The function computes the Jaccard index between all pairs of cells.

jaccardIndex(object, i, by = NULL)

Arguments

object

An object of class QFeatures.

i

The index of the assay in object. The assay must contain an identification matrix, that is a matrix where an entry is TRUE if the value is observed and FALSE is the value is missing (see examples).

by

A vector of length equal to the number of columns in assay i that defines groups for which the Jaccard index should be computed separately. If missing, the Jaccard indices are computed for all airs of cells in the dataset.

Value

A data.frame with as many rows as pairs of cells and the following column(s):

jaccard: the computed Jaccard index

by: if by is not NULL, the group of the pair of cells for which the Jaccard index is computed.

Examples

data("scp1")

## Define the identification matrix
peps <- scp1[["peptides"]]
assay(peps) <- ifelse(is.na(assay(peps)), FALSE, TRUE)
scp1 <- addAssay(scp1, peps, "id")

## Compute Jaccard indices
jaccardIndex(scp1, "id")
#>         jaccard  by
#> 1   1.000000000 all
#> 2   1.000000000 all
#> 3   1.000000000 all
#> 4   1.000000000 all
#> 5   1.000000000 all
#> 6   1.000000000 all
#> 7   1.000000000 all
#> 8   1.000000000 all
#> 9   1.000000000 all
#> 10  1.000000000 all
#> 11  1.000000000 all
#> 12  1.000000000 all
#> 13  1.000000000 all
#> 14  1.000000000 all
#> 15  1.000000000 all
#> 16  1.000000000 all
#> 17  1.000000000 all
#> 18  1.000000000 all
#> 19  1.000000000 all
#> 20  1.000000000 all
#> 21  1.000000000 all
#> 22  1.000000000 all
#> 23  1.000000000 all
#> 24  1.000000000 all
#> 25  1.000000000 all
#> 26  1.000000000 all
#> 27  1.000000000 all
#> 28  1.000000000 all
#> 29  1.000000000 all
#> 30  1.000000000 all
#> 31  1.000000000 all
#> 32  1.000000000 all
#> 33  1.000000000 all
#> 34  1.000000000 all
#> 35  1.000000000 all
#> 36  1.000000000 all
#> 37  1.000000000 all
#> 38  1.000000000 all
#> 39  1.000000000 all
#> 40  1.000000000 all
#> 41  1.000000000 all
#> 42  1.000000000 all
#> 43  1.000000000 all
#> 44  1.000000000 all
#> 45  1.000000000 all
#> 46  1.000000000 all
#> 47  1.000000000 all
#> 48  1.000000000 all
#> 49  1.000000000 all
#> 50  1.000000000 all
#> 51  1.000000000 all
#> 52  1.000000000 all
#> 53  1.000000000 all
#> 54  1.000000000 all
#> 55  1.000000000 all
#> 56  0.030674847 all
#> 57  0.030674847 all
#> 58  0.030674847 all
#> 59  0.030674847 all
#> 60  0.030674847 all
#> 61  0.030674847 all
#> 62  0.030674847 all
#> 63  0.030674847 all
#> 64  0.030674847 all
#> 65  0.030674847 all
#> 66  0.030674847 all
#> 67  0.030674847 all
#> 68  0.030674847 all
#> 69  0.030674847 all
#> 70  0.030674847 all
#> 71  0.030674847 all
#> 72  0.030674847 all
#> 73  0.030674847 all
#> 74  0.030674847 all
#> 75  0.030674847 all
#> 76  0.030674847 all
#> 77  0.030674847 all
#> 78  1.000000000 all
#> 79  0.030674847 all
#> 80  0.030674847 all
#> 81  0.030674847 all
#> 82  0.030674847 all
#> 83  0.030674847 all
#> 84  0.030674847 all
#> 85  0.030674847 all
#> 86  0.030674847 all
#> 87  0.030674847 all
#> 88  0.030674847 all
#> 89  0.030674847 all
#> 90  1.000000000 all
#> 91  1.000000000 all
#> 92  0.030674847 all
#> 93  0.030674847 all
#> 94  0.030674847 all
#> 95  0.030674847 all
#> 96  0.030674847 all
#> 97  0.030674847 all
#> 98  0.030674847 all
#> 99  0.030674847 all
#> 100 0.030674847 all
#> 101 0.030674847 all
#> 102 0.030674847 all
#> 103 1.000000000 all
#> 104 1.000000000 all
#> 105 1.000000000 all
#> 106 0.030674847 all
#> 107 0.030674847 all
#> 108 0.030674847 all
#> 109 0.030674847 all
#> 110 0.030674847 all
#> 111 0.030674847 all
#> 112 0.030674847 all
#> 113 0.030674847 all
#> 114 0.030674847 all
#> 115 0.030674847 all
#> 116 0.030674847 all
#> 117 1.000000000 all
#> 118 1.000000000 all
#> 119 1.000000000 all
#> 120 1.000000000 all
#> 121 0.030674847 all
#> 122 0.030674847 all
#> 123 0.030674847 all
#> 124 0.030674847 all
#> 125 0.030674847 all
#> 126 0.030674847 all
#> 127 0.030674847 all
#> 128 0.030674847 all
#> 129 0.030674847 all
#> 130 0.030674847 all
#> 131 0.030674847 all
#> 132 1.000000000 all
#> 133 1.000000000 all
#> 134 1.000000000 all
#> 135 1.000000000 all
#> 136 1.000000000 all
#> 137 0.030674847 all
#> 138 0.030674847 all
#> 139 0.030674847 all
#> 140 0.030674847 all
#> 141 0.030674847 all
#> 142 0.030674847 all
#> 143 0.030674847 all
#> 144 0.030674847 all
#> 145 0.030674847 all
#> 146 0.030674847 all
#> 147 0.030674847 all
#> 148 1.000000000 all
#> 149 1.000000000 all
#> 150 1.000000000 all
#> 151 1.000000000 all
#> 152 1.000000000 all
#> 153 1.000000000 all
#> 154 0.030674847 all
#> 155 0.030674847 all
#> 156 0.030674847 all
#> 157 0.030674847 all
#> 158 0.030674847 all
#> 159 0.030674847 all
#> 160 0.030674847 all
#> 161 0.030674847 all
#> 162 0.030674847 all
#> 163 0.030674847 all
#> 164 0.030674847 all
#> 165 1.000000000 all
#> 166 1.000000000 all
#> 167 1.000000000 all
#> 168 1.000000000 all
#> 169 1.000000000 all
#> 170 1.000000000 all
#> 171 1.000000000 all
#> 172 0.030674847 all
#> 173 0.030674847 all
#> 174 0.030674847 all
#> 175 0.030674847 all
#> 176 0.030674847 all
#> 177 0.030674847 all
#> 178 0.030674847 all
#> 179 0.030674847 all
#> 180 0.030674847 all
#> 181 0.030674847 all
#> 182 0.030674847 all
#> 183 1.000000000 all
#> 184 1.000000000 all
#> 185 1.000000000 all
#> 186 1.000000000 all
#> 187 1.000000000 all
#> 188 1.000000000 all
#> 189 1.000000000 all
#> 190 1.000000000 all
#> 191 0.030674847 all
#> 192 0.030674847 all
#> 193 0.030674847 all
#> 194 0.030674847 all
#> 195 0.030674847 all
#> 196 0.030674847 all
#> 197 0.030674847 all
#> 198 0.030674847 all
#> 199 0.030674847 all
#> 200 0.030674847 all
#> 201 0.030674847 all
#> 202 1.000000000 all
#> 203 1.000000000 all
#> 204 1.000000000 all
#> 205 1.000000000 all
#> 206 1.000000000 all
#> 207 1.000000000 all
#> 208 1.000000000 all
#> 209 1.000000000 all
#> 210 1.000000000 all
#> 211 0.030674847 all
#> 212 0.030674847 all
#> 213 0.030674847 all
#> 214 0.030674847 all
#> 215 0.030674847 all
#> 216 0.030674847 all
#> 217 0.030674847 all
#> 218 0.030674847 all
#> 219 0.030674847 all
#> 220 0.030674847 all
#> 221 0.030674847 all
#> 222 1.000000000 all
#> 223 1.000000000 all
#> 224 1.000000000 all
#> 225 1.000000000 all
#> 226 1.000000000 all
#> 227 1.000000000 all
#> 228 1.000000000 all
#> 229 1.000000000 all
#> 230 1.000000000 all
#> 231 1.000000000 all
#> 232 0.000000000 all
#> 233 0.000000000 all
#> 234 0.000000000 all
#> 235 0.000000000 all
#> 236 0.000000000 all
#> 237 0.000000000 all
#> 238 0.000000000 all
#> 239 0.000000000 all
#> 240 0.000000000 all
#> 241 0.000000000 all
#> 242 0.000000000 all
#> 243 0.002597403 all
#> 244 0.002597403 all
#> 245 0.002597403 all
#> 246 0.002597403 all
#> 247 0.002597403 all
#> 248 0.002597403 all
#> 249 0.002597403 all
#> 250 0.002597403 all
#> 251 0.002597403 all
#> 252 0.002597403 all
#> 253 0.002597403 all
#> 254 0.000000000 all
#> 255 0.000000000 all
#> 256 0.000000000 all
#> 257 0.000000000 all
#> 258 0.000000000 all
#> 259 0.000000000 all
#> 260 0.000000000 all
#> 261 0.000000000 all
#> 262 0.000000000 all
#> 263 0.000000000 all
#> 264 0.000000000 all
#> 265 0.002597403 all
#> 266 0.002597403 all
#> 267 0.002597403 all
#> 268 0.002597403 all
#> 269 0.002597403 all
#> 270 0.002597403 all
#> 271 0.002597403 all
#> 272 0.002597403 all
#> 273 0.002597403 all
#> 274 0.002597403 all
#> 275 0.002597403 all
#> 276 1.000000000 all
#> 277 0.000000000 all
#> 278 0.000000000 all
#> 279 0.000000000 all
#> 280 0.000000000 all
#> 281 0.000000000 all
#> 282 0.000000000 all
#> 283 0.000000000 all
#> 284 0.000000000 all
#> 285 0.000000000 all
#> 286 0.000000000 all
#> 287 0.000000000 all
#> 288 0.002597403 all
#> 289 0.002597403 all
#> 290 0.002597403 all
#> 291 0.002597403 all
#> 292 0.002597403 all
#> 293 0.002597403 all
#> 294 0.002597403 all
#> 295 0.002597403 all
#> 296 0.002597403 all
#> 297 0.002597403 all
#> 298 0.002597403 all
#> 299 1.000000000 all
#> 300 1.000000000 all
#> 301 0.000000000 all
#> 302 0.000000000 all
#> 303 0.000000000 all
#> 304 0.000000000 all
#> 305 0.000000000 all
#> 306 0.000000000 all
#> 307 0.000000000 all
#> 308 0.000000000 all
#> 309 0.000000000 all
#> 310 0.000000000 all
#> 311 0.000000000 all
#> 312 0.002597403 all
#> 313 0.002597403 all
#> 314 0.002597403 all
#> 315 0.002597403 all
#> 316 0.002597403 all
#> 317 0.002597403 all
#> 318 0.002597403 all
#> 319 0.002597403 all
#> 320 0.002597403 all
#> 321 0.002597403 all
#> 322 0.002597403 all
#> 323 1.000000000 all
#> 324 1.000000000 all
#> 325 1.000000000 all
#> 326 0.000000000 all
#> 327 0.000000000 all
#> 328 0.000000000 all
#> 329 0.000000000 all
#> 330 0.000000000 all
#> 331 0.000000000 all
#> 332 0.000000000 all
#> 333 0.000000000 all
#> 334 0.000000000 all
#> 335 0.000000000 all
#> 336 0.000000000 all
#> 337 0.002597403 all
#> 338 0.002597403 all
#> 339 0.002597403 all
#> 340 0.002597403 all
#> 341 0.002597403 all
#> 342 0.002597403 all
#> 343 0.002597403 all
#> 344 0.002597403 all
#> 345 0.002597403 all
#> 346 0.002597403 all
#> 347 0.002597403 all
#> 348 1.000000000 all
#> 349 1.000000000 all
#> 350 1.000000000 all
#> 351 1.000000000 all
#> 352 0.000000000 all
#> 353 0.000000000 all
#> 354 0.000000000 all
#> 355 0.000000000 all
#> 356 0.000000000 all
#> 357 0.000000000 all
#> 358 0.000000000 all
#> 359 0.000000000 all
#> 360 0.000000000 all
#> 361 0.000000000 all
#> 362 0.000000000 all
#> 363 0.002597403 all
#> 364 0.002597403 all
#> 365 0.002597403 all
#> 366 0.002597403 all
#> 367 0.002597403 all
#> 368 0.002597403 all
#> 369 0.002597403 all
#> 370 0.002597403 all
#> 371 0.002597403 all
#> 372 0.002597403 all
#> 373 0.002597403 all
#> 374 1.000000000 all
#> 375 1.000000000 all
#> 376 1.000000000 all
#> 377 1.000000000 all
#> 378 1.000000000 all
#> 379 0.000000000 all
#> 380 0.000000000 all
#> 381 0.000000000 all
#> 382 0.000000000 all
#> 383 0.000000000 all
#> 384 0.000000000 all
#> 385 0.000000000 all
#> 386 0.000000000 all
#> 387 0.000000000 all
#> 388 0.000000000 all
#> 389 0.000000000 all
#> 390 0.002597403 all
#> 391 0.002597403 all
#> 392 0.002597403 all
#> 393 0.002597403 all
#> 394 0.002597403 all
#> 395 0.002597403 all
#> 396 0.002597403 all
#> 397 0.002597403 all
#> 398 0.002597403 all
#> 399 0.002597403 all
#> 400 0.002597403 all
#> 401 1.000000000 all
#> 402 1.000000000 all
#> 403 1.000000000 all
#> 404 1.000000000 all
#> 405 1.000000000 all
#> 406 1.000000000 all
#> 407 0.000000000 all
#> 408 0.000000000 all
#> 409 0.000000000 all
#> 410 0.000000000 all
#> 411 0.000000000 all
#> 412 0.000000000 all
#> 413 0.000000000 all
#> 414 0.000000000 all
#> 415 0.000000000 all
#> 416 0.000000000 all
#> 417 0.000000000 all
#> 418 0.002597403 all
#> 419 0.002597403 all
#> 420 0.002597403 all
#> 421 0.002597403 all
#> 422 0.002597403 all
#> 423 0.002597403 all
#> 424 0.002597403 all
#> 425 0.002597403 all
#> 426 0.002597403 all
#> 427 0.002597403 all
#> 428 0.002597403 all
#> 429 1.000000000 all
#> 430 1.000000000 all
#> 431 1.000000000 all
#> 432 1.000000000 all
#> 433 1.000000000 all
#> 434 1.000000000 all
#> 435 1.000000000 all
#> 436 0.000000000 all
#> 437 0.000000000 all
#> 438 0.000000000 all
#> 439 0.000000000 all
#> 440 0.000000000 all
#> 441 0.000000000 all
#> 442 0.000000000 all
#> 443 0.000000000 all
#> 444 0.000000000 all
#> 445 0.000000000 all
#> 446 0.000000000 all
#> 447 0.002597403 all
#> 448 0.002597403 all
#> 449 0.002597403 all
#> 450 0.002597403 all
#> 451 0.002597403 all
#> 452 0.002597403 all
#> 453 0.002597403 all
#> 454 0.002597403 all
#> 455 0.002597403 all
#> 456 0.002597403 all
#> 457 0.002597403 all
#> 458 1.000000000 all
#> 459 1.000000000 all
#> 460 1.000000000 all
#> 461 1.000000000 all
#> 462 1.000000000 all
#> 463 1.000000000 all
#> 464 1.000000000 all
#> 465 1.000000000 all
#> 466 0.000000000 all
#> 467 0.000000000 all
#> 468 0.000000000 all
#> 469 0.000000000 all
#> 470 0.000000000 all
#> 471 0.000000000 all
#> 472 0.000000000 all
#> 473 0.000000000 all
#> 474 0.000000000 all
#> 475 0.000000000 all
#> 476 0.000000000 all
#> 477 0.002597403 all
#> 478 0.002597403 all
#> 479 0.002597403 all
#> 480 0.002597403 all
#> 481 0.002597403 all
#> 482 0.002597403 all
#> 483 0.002597403 all
#> 484 0.002597403 all
#> 485 0.002597403 all
#> 486 0.002597403 all
#> 487 0.002597403 all
#> 488 1.000000000 all
#> 489 1.000000000 all
#> 490 1.000000000 all
#> 491 1.000000000 all
#> 492 1.000000000 all
#> 493 1.000000000 all
#> 494 1.000000000 all
#> 495 1.000000000 all
#> 496 1.000000000 all
#> 497 0.000000000 all
#> 498 0.000000000 all
#> 499 0.000000000 all
#> 500 0.000000000 all
#> 501 0.000000000 all
#> 502 0.000000000 all
#> 503 0.000000000 all
#> 504 0.000000000 all
#> 505 0.000000000 all
#> 506 0.000000000 all
#> 507 0.000000000 all
#> 508 0.002597403 all
#> 509 0.002597403 all
#> 510 0.002597403 all
#> 511 0.002597403 all
#> 512 0.002597403 all
#> 513 0.002597403 all
#> 514 0.002597403 all
#> 515 0.002597403 all
#> 516 0.002597403 all
#> 517 0.002597403 all
#> 518 0.002597403 all
#> 519 1.000000000 all
#> 520 1.000000000 all
#> 521 1.000000000 all
#> 522 1.000000000 all
#> 523 1.000000000 all
#> 524 1.000000000 all
#> 525 1.000000000 all
#> 526 1.000000000 all
#> 527 1.000000000 all
#> 528 1.000000000 all
#> 529 0.000000000 all
#> 530 0.000000000 all
#> 531 0.000000000 all
#> 532 0.000000000 all
#> 533 0.000000000 all
#> 534 0.000000000 all
#> 535 0.000000000 all
#> 536 0.000000000 all
#> 537 0.000000000 all
#> 538 0.000000000 all
#> 539 0.000000000 all
#> 540 0.002597403 all
#> 541 0.002597403 all
#> 542 0.002597403 all
#> 543 0.002597403 all
#> 544 0.002597403 all
#> 545 0.002597403 all
#> 546 0.002597403 all
#> 547 0.002597403 all
#> 548 0.002597403 all
#> 549 0.002597403 all
#> 550 0.002597403 all
#> 551 1.000000000 all
#> 552 1.000000000 all
#> 553 1.000000000 all
#> 554 1.000000000 all
#> 555 1.000000000 all
#> 556 1.000000000 all
#> 557 1.000000000 all
#> 558 1.000000000 all
#> 559 1.000000000 all
#> 560 1.000000000 all
#> 561 1.000000000 all
#> 562 0.000000000 all
#> 563 0.000000000 all
#> 564 0.000000000 all
#> 565 0.000000000 all
#> 566 0.000000000 all
#> 567 0.000000000 all
#> 568 0.000000000 all
#> 569 0.000000000 all
#> 570 0.000000000 all
#> 571 0.000000000 all
#> 572 0.000000000 all
#> 573 0.002597403 all
#> 574 0.002597403 all
#> 575 0.002597403 all
#> 576 0.002597403 all
#> 577 0.002597403 all
#> 578 0.002597403 all
#> 579 0.002597403 all
#> 580 0.002597403 all
#> 581 0.002597403 all
#> 582 0.002597403 all
#> 583 0.002597403 all
#> 584 1.000000000 all
#> 585 1.000000000 all
#> 586 1.000000000 all
#> 587 1.000000000 all
#> 588 1.000000000 all
#> 589 1.000000000 all
#> 590 1.000000000 all
#> 591 1.000000000 all
#> 592 1.000000000 all
#> 593 1.000000000 all
#> 594 1.000000000 all
#> 595 1.000000000 all
#> 596 0.000000000 all
#> 597 0.000000000 all
#> 598 0.000000000 all
#> 599 0.000000000 all
#> 600 0.000000000 all
#> 601 0.000000000 all
#> 602 0.000000000 all
#> 603 0.000000000 all
#> 604 0.000000000 all
#> 605 0.000000000 all
#> 606 0.000000000 all
#> 607 0.002597403 all
#> 608 0.002597403 all
#> 609 0.002597403 all
#> 610 0.002597403 all
#> 611 0.002597403 all
#> 612 0.002597403 all
#> 613 0.002597403 all
#> 614 0.002597403 all
#> 615 0.002597403 all
#> 616 0.002597403 all
#> 617 0.002597403 all
#> 618 1.000000000 all
#> 619 1.000000000 all
#> 620 1.000000000 all
#> 621 1.000000000 all
#> 622 1.000000000 all
#> 623 1.000000000 all
#> 624 1.000000000 all
#> 625 1.000000000 all
#> 626 1.000000000 all
#> 627 1.000000000 all
#> 628 1.000000000 all
#> 629 1.000000000 all
#> 630 1.000000000 all
#> 631 0.000000000 all
#> 632 0.000000000 all
#> 633 0.000000000 all
#> 634 0.000000000 all
#> 635 0.000000000 all
#> 636 0.000000000 all
#> 637 0.000000000 all
#> 638 0.000000000 all
#> 639 0.000000000 all
#> 640 0.000000000 all
#> 641 0.000000000 all
#> 642 0.002597403 all
#> 643 0.002597403 all
#> 644 0.002597403 all
#> 645 0.002597403 all
#> 646 0.002597403 all
#> 647 0.002597403 all
#> 648 0.002597403 all
#> 649 0.002597403 all
#> 650 0.002597403 all
#> 651 0.002597403 all
#> 652 0.002597403 all
#> 653 1.000000000 all
#> 654 1.000000000 all
#> 655 1.000000000 all
#> 656 1.000000000 all
#> 657 1.000000000 all
#> 658 1.000000000 all
#> 659 1.000000000 all
#> 660 1.000000000 all
#> 661 1.000000000 all
#> 662 1.000000000 all
#> 663 1.000000000 all
#> 664 1.000000000 all
#> 665 1.000000000 all
#> 666 1.000000000 all
#> 667 0.000000000 all
#> 668 0.000000000 all
#> 669 0.000000000 all
#> 670 0.000000000 all
#> 671 0.000000000 all
#> 672 0.000000000 all
#> 673 0.000000000 all
#> 674 0.000000000 all
#> 675 0.000000000 all
#> 676 0.000000000 all
#> 677 0.000000000 all
#> 678 0.002597403 all
#> 679 0.002597403 all
#> 680 0.002597403 all
#> 681 0.002597403 all
#> 682 0.002597403 all
#> 683 0.002597403 all
#> 684 0.002597403 all
#> 685 0.002597403 all
#> 686 0.002597403 all
#> 687 0.002597403 all
#> 688 0.002597403 all
#> 689 1.000000000 all
#> 690 1.000000000 all
#> 691 1.000000000 all
#> 692 1.000000000 all
#> 693 1.000000000 all
#> 694 1.000000000 all
#> 695 1.000000000 all
#> 696 1.000000000 all
#> 697 1.000000000 all
#> 698 1.000000000 all
#> 699 1.000000000 all
#> 700 1.000000000 all
#> 701 1.000000000 all
#> 702 1.000000000 all
#> 703 1.000000000 all
## Compute Jaccard indices by sample type
jaccardIndex(scp1, "id", scp1$SampleType)
#>         jaccard         by
#> 1   0.030674847    Carrier
#> 2   0.000000000    Carrier
#> 3   0.002597403    Carrier
#> 4   0.030674847  Reference
#> 5   0.000000000  Reference
#> 6   0.002597403  Reference
#> 7   0.030674847     Unused
#> 8   0.000000000     Unused
#> 9   0.002597403     Unused
#> 10  0.000000000     Unused
#> 11  0.002597403     Unused
#> 12  1.000000000     Unused
#> 13  1.000000000   Monocyte
#> 14  0.030674847   Monocyte
#> 15  0.030674847   Monocyte
#> 16  0.000000000   Monocyte
#> 17  0.000000000   Monocyte
#> 18  0.002597403   Monocyte
#> 19  0.000000000   Monocyte
#> 20  0.000000000   Monocyte
#> 21  0.002597403   Monocyte
#> 22  1.000000000   Monocyte
#> 23  0.000000000      Blank
#> 24  0.000000000      Blank
#> 25  1.000000000      Blank
#> 26  1.000000000 Macrophage
#> 27  1.000000000 Macrophage
#> 28  1.000000000 Macrophage
#> 29  1.000000000 Macrophage
#> 30  1.000000000 Macrophage
#> 31  1.000000000 Macrophage
#> 32  1.000000000 Macrophage
#> 33  1.000000000 Macrophage
#> 34  1.000000000 Macrophage
#> 35  1.000000000 Macrophage
#> 36  0.030674847 Macrophage
#> 37  0.030674847 Macrophage
#> 38  0.030674847 Macrophage
#> 39  0.030674847 Macrophage
#> 40  0.030674847 Macrophage
#> 41  0.030674847 Macrophage
#> 42  0.030674847 Macrophage
#> 43  0.030674847 Macrophage
#> 44  0.030674847 Macrophage
#> 45  0.030674847 Macrophage
#> 46  1.000000000 Macrophage
#> 47  0.030674847 Macrophage
#> 48  0.030674847 Macrophage
#> 49  0.030674847 Macrophage
#> 50  0.030674847 Macrophage
#> 51  0.030674847 Macrophage
#> 52  1.000000000 Macrophage
#> 53  1.000000000 Macrophage
#> 54  0.030674847 Macrophage
#> 55  0.030674847 Macrophage
#> 56  0.030674847 Macrophage
#> 57  0.030674847 Macrophage
#> 58  0.030674847 Macrophage
#> 59  1.000000000 Macrophage
#> 60  1.000000000 Macrophage
#> 61  1.000000000 Macrophage
#> 62  0.030674847 Macrophage
#> 63  0.030674847 Macrophage
#> 64  0.030674847 Macrophage
#> 65  0.030674847 Macrophage
#> 66  0.030674847 Macrophage
#> 67  1.000000000 Macrophage
#> 68  1.000000000 Macrophage
#> 69  1.000000000 Macrophage
#> 70  1.000000000 Macrophage
#> 71  0.030674847 Macrophage
#> 72  0.030674847 Macrophage
#> 73  0.030674847 Macrophage
#> 74  0.030674847 Macrophage
#> 75  0.030674847 Macrophage
#> 76  1.000000000 Macrophage
#> 77  1.000000000 Macrophage
#> 78  1.000000000 Macrophage
#> 79  1.000000000 Macrophage
#> 80  1.000000000 Macrophage
#> 81  0.030674847 Macrophage
#> 82  0.030674847 Macrophage
#> 83  0.030674847 Macrophage
#> 84  0.030674847 Macrophage
#> 85  0.030674847 Macrophage
#> 86  1.000000000 Macrophage
#> 87  1.000000000 Macrophage
#> 88  1.000000000 Macrophage
#> 89  1.000000000 Macrophage
#> 90  1.000000000 Macrophage
#> 91  1.000000000 Macrophage
#> 92  0.000000000 Macrophage
#> 93  0.000000000 Macrophage
#> 94  0.000000000 Macrophage
#> 95  0.000000000 Macrophage
#> 96  0.000000000 Macrophage
#> 97  0.002597403 Macrophage
#> 98  0.002597403 Macrophage
#> 99  0.002597403 Macrophage
#> 100 0.002597403 Macrophage
#> 101 0.002597403 Macrophage
#> 102 0.002597403 Macrophage
#> 103 0.002597403 Macrophage
#> 104 0.000000000 Macrophage
#> 105 0.000000000 Macrophage
#> 106 0.000000000 Macrophage
#> 107 0.000000000 Macrophage
#> 108 0.000000000 Macrophage
#> 109 0.002597403 Macrophage
#> 110 0.002597403 Macrophage
#> 111 0.002597403 Macrophage
#> 112 0.002597403 Macrophage
#> 113 0.002597403 Macrophage
#> 114 0.002597403 Macrophage
#> 115 0.002597403 Macrophage
#> 116 1.000000000 Macrophage
#> 117 0.000000000 Macrophage
#> 118 0.000000000 Macrophage
#> 119 0.000000000 Macrophage
#> 120 0.000000000 Macrophage
#> 121 0.000000000 Macrophage
#> 122 0.002597403 Macrophage
#> 123 0.002597403 Macrophage
#> 124 0.002597403 Macrophage
#> 125 0.002597403 Macrophage
#> 126 0.002597403 Macrophage
#> 127 0.002597403 Macrophage
#> 128 0.002597403 Macrophage
#> 129 1.000000000 Macrophage
#> 130 1.000000000 Macrophage
#> 131 0.000000000 Macrophage
#> 132 0.000000000 Macrophage
#> 133 0.000000000 Macrophage
#> 134 0.000000000 Macrophage
#> 135 0.000000000 Macrophage
#> 136 0.002597403 Macrophage
#> 137 0.002597403 Macrophage
#> 138 0.002597403 Macrophage
#> 139 0.002597403 Macrophage
#> 140 0.002597403 Macrophage
#> 141 0.002597403 Macrophage
#> 142 0.002597403 Macrophage
#> 143 1.000000000 Macrophage
#> 144 1.000000000 Macrophage
#> 145 1.000000000 Macrophage
#> 146 0.000000000 Macrophage
#> 147 0.000000000 Macrophage
#> 148 0.000000000 Macrophage
#> 149 0.000000000 Macrophage
#> 150 0.000000000 Macrophage
#> 151 0.002597403 Macrophage
#> 152 0.002597403 Macrophage
#> 153 0.002597403 Macrophage
#> 154 0.002597403 Macrophage
#> 155 0.002597403 Macrophage
#> 156 0.002597403 Macrophage
#> 157 0.002597403 Macrophage
#> 158 1.000000000 Macrophage
#> 159 1.000000000 Macrophage
#> 160 1.000000000 Macrophage
#> 161 1.000000000 Macrophage
#> 162 0.000000000 Macrophage
#> 163 0.000000000 Macrophage
#> 164 0.000000000 Macrophage
#> 165 0.000000000 Macrophage
#> 166 0.000000000 Macrophage
#> 167 0.002597403 Macrophage
#> 168 0.002597403 Macrophage
#> 169 0.002597403 Macrophage
#> 170 0.002597403 Macrophage
#> 171 0.002597403 Macrophage
#> 172 0.002597403 Macrophage
#> 173 0.002597403 Macrophage
#> 174 1.000000000 Macrophage
#> 175 1.000000000 Macrophage
#> 176 1.000000000 Macrophage
#> 177 1.000000000 Macrophage
#> 178 1.000000000 Macrophage
#> 179 0.000000000 Macrophage
#> 180 0.000000000 Macrophage
#> 181 0.000000000 Macrophage
#> 182 0.000000000 Macrophage
#> 183 0.000000000 Macrophage
#> 184 0.002597403 Macrophage
#> 185 0.002597403 Macrophage
#> 186 0.002597403 Macrophage
#> 187 0.002597403 Macrophage
#> 188 0.002597403 Macrophage
#> 189 0.002597403 Macrophage
#> 190 0.002597403 Macrophage
#> 191 1.000000000 Macrophage
#> 192 1.000000000 Macrophage
#> 193 1.000000000 Macrophage
#> 194 1.000000000 Macrophage
#> 195 1.000000000 Macrophage
#> 196 1.000000000 Macrophage
#> 197 0.000000000 Macrophage
#> 198 0.000000000 Macrophage
#> 199 0.000000000 Macrophage
#> 200 0.000000000 Macrophage
#> 201 0.000000000 Macrophage
#> 202 0.002597403 Macrophage
#> 203 0.002597403 Macrophage
#> 204 0.002597403 Macrophage
#> 205 0.002597403 Macrophage
#> 206 0.002597403 Macrophage
#> 207 0.002597403 Macrophage
#> 208 0.002597403 Macrophage
#> 209 1.000000000 Macrophage
#> 210 1.000000000 Macrophage
#> 211 1.000000000 Macrophage
#> 212 1.000000000 Macrophage
#> 213 1.000000000 Macrophage
#> 214 1.000000000 Macrophage
#> 215 1.000000000 Macrophage