Read DIA-NN output as a QFeatures objects for single-cell proteomics data

This function takes the output tables from DIA-NN and converts them into a QFeatures object using the scp framework.

readSCPfromDIANN(
  colData,
  reportData,
  extractedData = NULL,
  ecol = "MS1.Area",
  multiplexing = "none",
  ...
)

Arguments

colData

A data.frame or any object that can be coerced to a data.frame. colData is expected to contains all the sample annotations. We require the table to contain a column called File.Name that links to the File.Name in the DIA-NN report table. If multiplexing = "mTRAQ", we require a second column called Label that links the label to the sample (the labels identified by DIA-NN can be retrieved from Modified Sequence column in the report table).

reportData

A data.frame or any object that can be coerced to a data.frame that contains the data from the Report.tsv file generated by DIA-NN.

extractedData

A data.frame or any object that can be coerced to a data.frame that contains the data from the *_ms1_extracted.tsv file generated by DIA-NN. This argument is optional and is only applicable for mulitplixed experiments

ecol

A character(1) indicating which column in reportData contains the quantitative information.

multiplexing

A character(1) indicating the type of multiplexing used in the experiment. Provide "none" if the experiment is label-free (default). Available options are: "mTRAQ".

...

Further arguments passed to readSCP()

Value

An instance of class QFeatures. The expression data of each acquisition run is stored in a separate assay as a SingleCellExperiment object.