Convert a seurat object to an anndata object using reticulate

Convert a seurat object to an anndata object using reticulate

Usage

srt_to_adata(
  srt,
  features = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  convert_tools = FALSE,
  convert_misc = FALSE,
  verbose = TRUE
)

Arguments

srt

A Seurat object.

assay_X

Assays to convert as X(main data matrix) in anndata object.

assay_layers

Assays to convert as layers in anndata object.

verbose

Value

A anndata object.

Examples

data("pancreas_sub")
adata <- srt_to_adata(pancreas_sub)
#> 'misc' slot is not converted.
#> 'tools' slot is not converted.
adata
#> AnnData object with n_obs × n_vars = 1000 × 15958
#>     obs: 'orig.ident', 'nCount_RNA', 'nFeature_RNA', 'S_score', 'G2M_score', 'nCount_spliced', 'nFeature_spliced', 'nCount_unspliced', 'nFeature_unspliced', 'CellType', 'SubCellType', 'Phase'
#>     var: 'features', 'highly_variable_genes'
#>     obsm: 'PCA', 'UMAP'
#>     layers: 'spliced', 'unspliced'

### Or save as an h5ad file or a loom file
# adata$write_h5ad("pancreas_sub.h5ad")
# adata$write_loom("pancreas_sub.loom", write_obsm_varm = TRUE)