Transfer the colData to an Assay

The function transfers the colData from a QFeatures object to one of the assays it contains. The transfered data is bound to the existing colData of the target assay.

transferColDataToAssay(obj, i)

Arguments

obj
A QFeatures object

i
A numeric(1) or character(1) indicating which assay to transfer the colData to.

Value

A QFeatures object

Examples

data("scp1")
colData(scp1[["peptides"]])

#> DataFrame with 38 rows and 0 columns

scp1 <- transferColDataToAssay(scp1, i = "peptides")
colData(scp1[["peptides"]])

#> DataFrame with 38 rows and 7 columns
#>                                             Set     Channel SampleAnnotation
#>                                     <character> <character>      <character>
#> 190321S_LCA10_X_FP97AG_RI1        190321S_LC...         RI1    carrier_mi...
#> 190321S_LCA10_X_FP97AG_RI2        190321S_LC...         RI2             norm
#> 190321S_LCA10_X_FP97AG_RI3        190321S_LC...         RI3           unused
#> 190321S_LCA10_X_FP97AG_RI4        190321S_LC...         RI4            sc_m0
#> 190321S_LCA10_X_FP97AG_RI5        190321S_LC...         RI5             sc_u
#> ...                                         ...         ...              ...
#> 190914S_LCB3_X_16plex_Set_21_RI12 190914S_LC...        RI12            sc_m0
#> 190914S_LCB3_X_16plex_Set_21_RI13 190914S_LC...        RI13            sc_m0
#> 190914S_LCB3_X_16plex_Set_21_RI14 190914S_LC...        RI14            sc_m0
#> 190914S_LCB3_X_16plex_Set_21_RI15 190914S_LC...        RI15            sc_m0
#> 190914S_LCB3_X_16plex_Set_21_RI16 190914S_LC...        RI16            sc_m0
#>                                    SampleType     lcbatch     sortday
#>                                   <character> <character> <character>
#> 190321S_LCA10_X_FP97AG_RI1            Carrier       LCA10          s8
#> 190321S_LCA10_X_FP97AG_RI2          Reference       LCA10          s8
#> 190321S_LCA10_X_FP97AG_RI3             Unused       LCA10          s8
#> 190321S_LCA10_X_FP97AG_RI4         Macrophage       LCA10          s8
#> 190321S_LCA10_X_FP97AG_RI5           Monocyte       LCA10          s8
#> ...                                       ...         ...         ...
#> 190914S_LCB3_X_16plex_Set_21_RI12  Macrophage        LCB3          s9
#> 190914S_LCB3_X_16plex_Set_21_RI13  Macrophage        LCB3          s9
#> 190914S_LCB3_X_16plex_Set_21_RI14  Macrophage        LCB3          s9
#> 190914S_LCB3_X_16plex_Set_21_RI15  Macrophage        LCB3          s9
#> 190914S_LCB3_X_16plex_Set_21_RI16  Macrophage        LCB3          s9
#>                                        digest
#>                                   <character>
#> 190321S_LCA10_X_FP97AG_RI1                  Q
#> 190321S_LCA10_X_FP97AG_RI2                  Q
#> 190321S_LCA10_X_FP97AG_RI3                  Q
#> 190321S_LCA10_X_FP97AG_RI4                  Q
#> 190321S_LCA10_X_FP97AG_RI5                  Q
#> ...                                       ...
#> 190914S_LCB3_X_16plex_Set_21_RI12           R
#> 190914S_LCB3_X_16plex_Set_21_RI13           R
#> 190914S_LCB3_X_16plex_Set_21_RI14           R
#> 190914S_LCB3_X_16plex_Set_21_RI15           R
#> 190914S_LCB3_X_16plex_Set_21_RI16           R