Prefetch cycle gene

Based on the human cell cycle genes, the cell cycle genes of the corresponding species were captured by homologous gene conversion.

Usage

CC_GenePrefetch(
  species = "Homo_sapiens",
  Ensembl_version = 103,
  mirror = NULL,
  max_tries = 5,
  use_cached_gene = TRUE
)

Arguments

species

Latin names for animals,i.e., "Homo_sapiens", "Mus_musculus"

Ensembl_version

Ensembl database version. If NULL, use the current release version.

mirror

Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.

max_tries

The maximum number of attempts to connect with the BioMart service.

use_cached_gene

Whether to use previously cached cell cycle gene conversion results for the species.

Value

A list of S-phase and G2M-phase genes.

GeneConvert

Examples

ccgenes <- CC_GenePrefetch("Homo_sapiens")
str(ccgenes)
#> List of 3
#>  $ res: NULL
#>  $ S  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...
ccgenes <- CC_GenePrefetch("Mus_musculus")
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connecting to the biomart...
#> Searching the dataset hsapiens ...
#> Connecting to the dataset hsapiens_gene_ensembl ...
#> Converting the geneIDs...
#> 97 genes mapped with ensembl_symbol
#> ==============================
#> 97 genes mapped
#> 0 genes unmapped
#> ==============================
str(ccgenes)
#> List of 3
#>  $ res:List of 7
#>   ..$ geneID_res     :'data.frame':	99 obs. of  4 variables:
#>   .. ..$ from_IDtype: chr [1:99] "ensembl_symbol" "ensembl_symbol" "ensembl_symbol" "ensembl_symbol" ...
#>   .. ..$ from_geneID: chr [1:99] "NCAPD2" "ANLN" "UBR7" "TACC3" ...
#>   .. ..$ to_IDtype  : chr [1:99] "symbol" "symbol" "symbol" "symbol" ...
#>   .. ..$ to_geneID  : chr [1:99] "Ncapd2" "Anln" "Ubr7" "Tacc3" ...
#>   ..$ geneID_collapse:'data.frame':	94 obs. of  2 variables:
#>   .. ..$ from_geneID: chr [1:94] "ANLN" "ANP32E" "AURKA" "AURKB" ...
#>   .. ..$ symbol     :List of 94
#>   .. .. ..$ : chr "Anln"
#>   .. .. ..$ : chr "Anp32e"
#>   .. .. ..$ : chr "Aurka"
#>   .. .. ..$ : chr "Aurkb"
#>   .. .. ..$ : chr "Birc5"
#>   .. .. ..$ : chr "Blm"
#>   .. .. ..$ : chr "Bub1"
#>   .. .. ..$ : chr "Casp8ap2"
#>   .. .. ..$ : chr "Cbx5"
#>   .. .. ..$ : chr "Ccnb2"
#>   .. .. ..$ : chr "Ccne2"
#>   .. .. ..$ : chr "Cdc20"
#>   .. .. ..$ : chr "Cdc25c"
#>   .. .. ..$ : chr "Cdc45"
#>   .. .. ..$ : chr "Cdc6"
#>   .. .. ..$ : chr "Cdca2"
#>   .. .. ..$ : chr "Cdca3"
#>   .. .. ..$ : chr "Cdca7"
#>   .. .. ..$ : chr "Cdca8"
#>   .. .. ..$ : chr "Cdk1"
#>   .. .. ..$ : chr "Cenpa"
#>   .. .. ..$ : chr "Cenpe"
#>   .. .. ..$ : chr "Cenpf"
#>   .. .. ..$ : chr "Cenpu"
#>   .. .. ..$ : chr "Chaf1b"
#>   .. .. ..$ : chr "Ckap2"
#>   .. .. ..$ : chr "Ckap2l"
#>   .. .. ..$ : chr "Ckap5"
#>   .. .. ..$ : chr "Cks1b"
#>   .. .. ..$ : chr "Cks2"
#>   .. .. ..$ : chr "Clspn"
#>   .. .. ..$ : chr "Ctcf"
#>   .. .. ..$ : chr "Dlgap5"
#>   .. .. ..$ : chr "Dscc1"
#>   .. .. ..$ : chr "Dtl"
#>   .. .. ..$ : chr "E2f8"
#>   .. .. ..$ : chr "Ect2"
#>   .. .. ..$ : chr "Exo1"
#>   .. .. ..$ : chr "Fen1"
#>   .. .. ..$ : chr "G2e3"
#>   .. .. ..$ : chr "Gas2l3"
#>   .. .. ..$ : chr "Gins2"
#>   .. .. ..$ : chr "Gmnn"
#>   .. .. ..$ : chr "Gtse1"
#>   .. .. ..$ : chr "Hells"
#>   .. .. ..$ : chr "Hjurp"
#>   .. .. ..$ : chr "Hmgb2"
#>   .. .. ..$ : chr "Hmmr"
#>   .. .. ..$ : chr "Kif11"
#>   .. .. ..$ : chr "Kif20b"
#>   .. .. ..$ : chr "Kif23"
#>   .. .. ..$ : chr "Kif2c"
#>   .. .. ..$ : chr "Lbr"
#>   .. .. ..$ : chr "Mcm4"
#>   .. .. ..$ : chr "Mcm5"
#>   .. .. ..$ : chr "Mcm6"
#>   .. .. ..$ : chr "Mcm7"
#>   .. .. ..$ : chr "Mki67"
#>   .. .. ..$ : chr "Mrpl36"
#>   .. .. ..$ : chr "Msh2"
#>   .. .. ..$ : chr "Nasp"
#>   .. .. ..$ : chr "Ncapd2"
#>   .. .. ..$ : chr "Ndc80"
#>   .. .. ..$ : chr "Nek2"
#>   .. .. ..$ : chr "Nuf2"
#>   .. .. ..$ : chr "Nusap1"
#>   .. .. ..$ : chr "Pcna"
#>   .. .. ..$ : chr "Pimreg"
#>   .. .. ..$ : chr "Pola1"
#>   .. .. ..$ : chr "Polr1b"
#>   .. .. ..$ : chr "Prim1"
#>   .. .. ..$ : chr "Psrc1"
#>   .. .. ..$ : chr "Rad51"
#>   .. .. ..$ : chr "Rad51ap1"
#>   .. .. ..$ : chr "Rangap1"
#>   .. .. ..$ : chr "Rfc2"
#>   .. .. ..$ : chr "Rrm1"
#>   .. .. ..$ : chr "Rrm2"
#>   .. .. ..$ : chr "Slbp"
#>   .. .. ..$ : chr "Smc4"
#>   .. .. ..$ : chr "Tacc3"
#>   .. .. ..$ : chr "Tipin"
#>   .. .. ..$ : chr "Tmpo"
#>   .. .. ..$ : chr "Top2a"
#>   .. .. ..$ : chr "Tpx2"
#>   .. .. ..$ : chr "Ttk"
#>   .. .. ..$ : chr "Tubb4b"
#>   .. .. ..$ : chr "Tyms"
#>   .. .. ..$ : chr "Ube2c"
#>   .. .. ..$ : chr "Ubr7"
#>   .. .. ..$ : chr "Uhrf1"
#>   .. .. ..$ : chr "Ung"
#>   .. .. ..$ : chr "Usp1"
#>   .. .. ..$ : chr "Wdr76"
#>   .. .. ..- attr(*, "class")= chr "AsIs"
#>   ..$ geneID_expand  :'data.frame':	94 obs. of  2 variables:
#>   .. ..$ from_geneID: chr [1:94] "ANLN" "ANP32E" "AURKA" "AURKB" ...
#>   .. ..$ symbol     : chr [1:94] "Anln" "Anp32e" "Aurka" "Aurkb" ...
#>   ..$ Ensembl_version: chr "103"
#>   ..$ Datasets       :'data.frame':	202 obs. of  3 variables:
#>   .. ..$ dataset    : 'AsIs' chr [1:202] "abrachyrhynchus_gene_ensembl" "acalliptera_gene_ensembl" "acarolinensis_gene_ensembl" "acchrysaetos_gene_ensembl" ...
#>   .. ..$ description: 'AsIs' chr [1:202] "Pink-footed goose genes (ASM259213v1)" "Eastern happy genes (fAstCal1.2)" "Anole lizard genes (AnoCar2.0)" "Golden eagle genes (bAquChr1.2)" ...
#>   .. ..$ version    : 'AsIs' chr [1:202] "ASM259213v1" "fAstCal1.2" "AnoCar2.0" "bAquChr1.2" ...
#>   ..$ Attributes     :'data.frame':	3183 obs. of  3 variables:
#>   .. ..$ name       : chr [1:3183] "ensembl_gene_id" "ensembl_gene_id_version" "ensembl_transcript_id" "ensembl_transcript_id_version" ...
#>   .. ..$ description: chr [1:3183] "Gene stable ID" "Gene stable ID version" "Transcript stable ID" "Transcript stable ID version" ...
#>   .. ..$ page       : chr [1:3183] "feature_page" "feature_page" "feature_page" "feature_page" ...
#>   ..$ geneID_unmapped: chr(0) 
#>  $ S  : chr [1:41] "Mcm5" "Pcna" "Tyms" "Fen1" ...
#>  $ G2M: chr [1:53] "Hmgb2" "Cdk1" "Nusap1" "Ube2c" ...