Prefetch cycle gene

Based on the human cell cycle genes, the cell cycle genes of the corresponding species were captured by homologous gene conversion.

Usage

CC_GenePrefetch(
  species = "Homo_sapiens",
  Ensembl_version = 103,
  mirror = NULL,
  max_tries = 5,
  use_cached_gene = TRUE
)

Arguments

species

Latin names for animals,i.e., "Homo_sapiens", "Mus_musculus"

Ensembl_version

Ensembl database version. If NULL, use the current release version.

mirror

Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.

max_tries

The maximum number of attempts to connect with the BioMart service.

use_cached_gene

Whether to use previously cached cell cycle gene conversion results for the species.

Value

A list of S-phase and G2M-phase genes.

GeneConvert

Examples

ccgenes <- CC_GenePrefetch("Homo_sapiens")
str(ccgenes)
#> List of 3
#>  $ res: NULL
#>  $ S  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...
ccgenes <- CC_GenePrefetch("Mus_musculus")
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connecting to the biomart...
#> Searching the dataset hsapiens ...
#> Connecting to the dataset hsapiens_gene_ensembl ...
#> Converting the geneIDs...
#> 97 genes mapped with ensembl_symbol
#> ==============================
#> 97 genes mapped
#> 0 genes unmapped
#> ==============================
#> Error in unnest(data = geneID_collapse, cols = colnames(geneID_collapse)[sapply(geneID_collapse,     class) %in% c("list", "AsIs")], keep_empty = FALSE): could not find function "unnest"
str(ccgenes)
#> List of 3
#>  $ res: NULL
#>  $ S  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...