3D-Dimensional reduction plot for cell classification visualization.

Plotting cell points on a reduced 3D space and coloring according to the groups of the cells.

Usage

ClassDimPlot3D(
  srt,
  group.by = "orig.ident",
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  palette = "Paired",
  palcolor = NULL,
  bg_color = "grey80",
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  lineages = NULL,
  lineage_palette = "Dark2",
  span = 0.75,
  arrow_reverse = FALSE,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)

Arguments

srt

A Seurat object.

group.by

Name of one or more metadata columns to group (color) cells by (for example, orig.ident).

reduction

Which dimensionality reduction to use.

palette

Name of palette to use. Default is "Paired".

bg_color

Color value for NA points.

pt.size

Point size for plotting.

cells.highlight

A vector of names of cells to highlight.

cols.highlight

A vector of colors to highlight the cells.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

sizes.highlight

Size of highlighted cells.

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

Examples

data(pancreas_sub)
pancreas_sub <- Standard_SCP(pancreas_sub)
#> [2022-12-11 14:10:52] Start Standard_SCP
#> [2022-12-11 14:10:52] Checking srtList... ...
#> Data 1/1 of the srtList is raw counts. Perform NormalizeData(logCPM) on the data ...
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from the existed HVF in srtList...
#> Number of available HVF: 2000
#> [2022-12-11 14:10:54] Finished checking.
#> [2022-12-11 14:10:54] Perform ScaleData on the data...
#> [2022-12-11 14:10:54] Perform linear dimension reduction (pca) on the data...
#> dims_estimate is 1:18 for 'pca'
#> [2022-12-11 14:10:55] Perform FindClusters (louvain) on the data...
#> [2022-12-11 14:10:55] Reorder clusters...
#> [2022-12-11 14:10:55] Perform nonlinear dimension reduction (umap) on the data...
#> [2022-12-11 14:11:02] Standard_SCP done
#> Elapsed time: 9.02 secs 
ClassDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")

pancreas_sub <- Standard_SCP(pancreas_sub)
#> [2022-12-11 14:11:02] Start Standard_SCP
#> [2022-12-11 14:11:02] Checking srtList... ...
#> Data 1/1 of the srtList has been log-normalized.
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from the existed HVF in srtList...
#> Number of available HVF: 2000
#> [2022-12-11 14:11:03] Finished checking.
#> [2022-12-11 14:11:03] Perform ScaleData on the data...
#> [2022-12-11 14:11:03] Perform linear dimension reduction (pca) on the data...
#> dims_estimate is 1:18 for 'pca'
#> [2022-12-11 14:11:04] Perform FindClusters (louvain) on the data...
#> [2022-12-11 14:11:04] Reorder clusters...
#> [2022-12-11 14:11:04] Perform nonlinear dimension reduction (umap) on the data...
#> [2022-12-11 14:11:10] Standard_SCP done
#> Elapsed time: 8.81 secs 
pancreas_sub <- RunSlingshot(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D")
ClassDimPlot3D(pancreas_sub, group.by = "SubCellType", reduction = "StandardpcaUMAP3D", lineages = paste0("Lineage", 1:2))