3D-Dimensional reduction plot for gene expression visualization.

Plotting cell points on a reduced 3D space and coloring according to the gene expression in the cells.

Usage

ExpDimPlot3D(
  srt,
  features = NULL,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  split.by = NULL,
  slot = "data",
  assay = "RNA",
  calculate_coexp = FALSE,
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)

Arguments

srt

A Seurat object.

features

Name of one or more metadata columns to group (color) cells by (for example, orig.ident).

reduction

Which dimensionality reduction to use.

split.by

Name of a metadata column to split plot by.

pt.size

Point size for plotting.

cells.highlight

A vector of names of cells to highlight.

cols.highlight

A vector of colors to highlight the cells.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

sizes.highlight

Size of highlighted cells.

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

force

Examples

data(pancreas_sub)
pancreas_sub <- Standard_SCP(pancreas_sub)
#> [2022-10-23 20:08:01] Start Standard_SCP
#> [2022-10-23 20:08:01] Checking srtList... ...
#> Data 1/1 of the srtList is raw counts. Perform NormalizeData(logCPM) on the data ...
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from the existed HVF in srtList...
#> [2022-10-23 20:08:03] Finished checking.
#> [2022-10-23 20:08:03] Perform ScaleData on the data...
#> [2022-10-23 20:08:03] Perform linear dimension reduction (pca) on the data...
#> dims_estimate is 1:18 for 'pca'
#> [2022-10-23 20:08:11] Perform FindClusters (louvain) on the data...
#> [2022-10-23 20:08:11] Reorder clusters...
#> [2022-10-23 20:08:11] Perform nonlinear dimension reduction (umap) on the data...
#> [2022-10-23 20:08:19] Standard_SCP done
#> Elapsed time: 17.57 secs 
ExpDimPlot3D(pancreas_sub, features = c("Ghrl", "Ins1", "Gcg", "Ins2"), reduction = "StandardpcaUMAP3D")