3D-Dimensional reduction plot for gene expression visualization.

Plotting cell points on a reduced 3D space and coloring according to the gene expression in the cells.

Usage

FeatureDimPlot3D(
  srt,
  features = NULL,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  split.by = NULL,
  slot = "data",
  assay = NULL,
  calculate_coexp = FALSE,
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)

Arguments

srt

A Seurat object.

features

Name of one or more metadata columns to group (color) cells by (for example, orig.ident).

reduction

Which dimensionality reduction to use.

split.by

Name of a metadata column to split plot by.

pt.size

Point size for plotting.

cells.highlight

A vector of names of cells to highlight.

cols.highlight

A vector of colors to highlight the cells.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

sizes.highlight

Size of highlighted cells.

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

Examples

data("pancreas_sub")
pancreas_sub <- Standard_SCP(pancreas_sub)
#> [2023-03-09 16:02:31] Start Standard_SCP
#> [2023-03-09 16:02:31] Checking srtList... ...
#> Data 1/1 of the srtList is raw_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from srtList...
#> Number of available HVF: 2000
#> [2023-03-09 16:02:32] Finished checking.
#> [2023-03-09 16:02:32] Perform ScaleData on the data...
#> [2023-03-09 16:02:32] Perform linear dimension reduction (pca) on the data...
#> [2023-03-09 16:02:33] Perform FindClusters (louvain) on the data...
#> [2023-03-09 16:02:33] Reorder clusters...
#> [2023-03-09 16:02:34] Perform nonlinear dimension reduction (umap) on the data...
#> Non-linear dimensionality reduction(umap) using Reduction(Standardpca, dims_range:1-50) as input
#> Non-linear dimensionality reduction(umap) using Reduction(Standardpca, dims_range:1-50) as input
#> [2023-03-09 16:02:42] Standard_SCP done
#> Elapsed time: 10.82 secs 
FeatureDimPlot3D(pancreas_sub, features = c("Ghrl", "Ins1", "Gcg", "Ins2"), reduction = "StandardpcaUMAP3D")