Gene ID conversion function using biomart

This function can convert different gene ID types within one species or bewteen two species using biomart service.

Usage

GeneConvert(
  geneID,
  geneID_from_IDtype = "symbol",
  geneID_to_IDtype = "entrez_id",
  species_from = "Homo_sapiens",
  species_to = NULL,
  Ensembl_version = 103,
  attempts = 5,
  mirror = NULL
)

Arguments

geneID

A vector of the geneID character.

geneID_from_IDtype

ID type of the geneID. e.g. "symbol", "ensembl_id", "entrez_id"

geneID_to_IDtype

ID type to convert to. e.g. "symbol", "ensembl_id", "entrez_id"

species_from

Latin names for animals of the input geneID.

species_to

Latin names for animals of the output geneID. e.g. "Homo_sapiens","Mus_musculus"

Ensembl_version

Ensembl database version. If NULL, use the current release version.

attempts

Number of attempts to connect with the biomart service.

mirror

Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.

Value

A list.

Examples

res <- GeneConvert(
  geneID = c("CDK1", "MKI67", "TOP2A", "AURKA", "CTCF"), geneID_from_IDtype = "symbol", geneID_to_IDtype = "entrez_id",
  species_from = "Homo_sapiens", species_to = "Mus_musculus", Ensembl_version = 103
)
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Ensembl site unresponsive, trying uswest mirror
#> Connect to the dataset mmusculus_gene_ensembl ...
#> Match and convert the geneID...
#> 5 genes mapped with ensembl_symbol
#> ==============================
#> 5 genes mapped
#> 0 genes unmapped
#> ==============================
str(res)
#> List of 7
#>  $ geneID_res     :'data.frame':	5 obs. of  4 variables:
#>   ..$ from_IDtype: chr [1:5] "ensembl_symbol" "ensembl_symbol" "ensembl_symbol" "ensembl_symbol" ...
#>   ..$ from_geneID: chr [1:5] "CTCF" "CDK1" "TOP2A" "AURKA" ...
#>   ..$ to_IDtype  : chr [1:5] "entrez_id" "entrez_id" "entrez_id" "entrez_id" ...
#>   ..$ to_geneID  : int [1:5] 13018 12534 21973 20878 17345
#>  $ geneID_collapse:'data.frame':	5 obs. of  2 variables:
#>   ..$ from_geneID: chr [1:5] "AURKA" "CDK1" "CTCF" "MKI67" ...
#>   ..$ entrez_id  :List of 5
#>   .. ..$ : int 20878
#>   .. ..$ : int 12534
#>   .. ..$ : int 13018
#>   .. ..$ : int 17345
#>   .. ..$ : int 21973
#>  $ geneID_expand  :'data.frame':	5 obs. of  2 variables:
#>   ..$ from_geneID: chr [1:5] "AURKA" "CDK1" "CTCF" "MKI67" ...
#>   ..$ entrez_id  : int [1:5] 20878 12534 13018 17345 21973
#>  $ Ensembl_version: chr "103"
#>  $ Datasets       :'data.frame':	202 obs. of  3 variables:
#>   ..$ dataset    : 'AsIs' chr [1:202] "abrachyrhynchus_gene_ensembl" "acalliptera_gene_ensembl" "acarolinensis_gene_ensembl" "acchrysaetos_gene_ensembl" ...
#>   ..$ description: 'AsIs' chr [1:202] "Pink-footed goose genes (ASM259213v1)" "Eastern happy genes (fAstCal1.2)" "Anole lizard genes (AnoCar2.0)" "Golden eagle genes (bAquChr1.2)" ...
#>   ..$ version    : 'AsIs' chr [1:202] "ASM259213v1" "fAstCal1.2" "AnoCar2.0" "bAquChr1.2" ...
#>  $ Attributes     :'data.frame':	3183 obs. of  3 variables:
#>   ..$ name       : chr [1:3183] "ensembl_gene_id" "ensembl_gene_id_version" "ensembl_transcript_id" "ensembl_transcript_id_version" ...
#>   ..$ description: chr [1:3183] "Gene stable ID" "Gene stable ID version" "Transcript stable ID" "Transcript stable ID version" ...
#>   ..$ page       : chr [1:3183] "feature_page" "feature_page" "feature_page" "feature_page" ...
#>  $ geneID_unmapped: chr(0)