Prepare the database for enrichment analysis

Prepare the database for enrichment analysis

Usage

PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"),
  db = c("GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome", "ProteinComplex",
    "DGI", "MP", "DO", "PFAM", "Chromosome", "GeneType", "Enzyme", "TF", "SP",
    "CellTalk", "CellChat"),
  db_IDtypes = c("symbol", "entrez_id", "ensembl_id"),
  db_version = "latest",
  db_update = FALSE,
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  custom_TERM2GENE = NULL,
  custom_TERM2NAME = NULL,
  custom_species = NULL,
  custom_IDtype = NULL,
  custom_version = NULL
)

Arguments

species

species

db

Enrichment database name.

db_update

Whether update the database to the latest version.

custom_version

Value

A list containing the database.

Examples

db_list <- PrepareDB(species = "Homo_sapiens", db = "GO_BP", db_update = TRUE)
#> Species: Homo_sapiens
#> Preparing database: GO_BP
#> Convert ID types for the database: GO_BP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Match and convert the geneID...
#> 18241 genes mapped with entrez_id
#> ==============================
#> 18241 genes mapped
#> 662 genes unmapped
#> ==============================
ListDB(species = "Homo_sapiens", db = "GO_BP")
#>                                                         identifier version                               comment
#> 6 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7 3.16.0 nterm:15975|Homo_sapiens-GO_BP
#>    timestamp                date         db_version            db_name
#> 6 1674031921 2023-01-18 08:52:00 3.16.0 nterm:15975 Homo_sapiens-GO_BP
#>                                                                    file
#> 6 /home/runner/.cache/R/R.cache/9a33e14dc99b256952de599e15c222ab.Rcache
head(db_list[["Homo_sapiens"]][["GO_BP"]][["TERM2GENE"]])
#>         Term entrez_id  symbol      ensembl_id
#> 1 GO:0000002     30968  STOML2 ENSG00000165283
#> 2 GO:0000002    201973 PRIMPOL ENSG00000164306
#> 3 GO:0000002      6742   SSBP1 ENSG00000106028
#> 4 GO:0000002      6742   SSBP1 ENSG00000262771
#> 5 GO:0000002      1763    DNA2 ENSG00000138346
#> 6 GO:0000002     11232   POLG2 ENSG00000256525

db_list <- PrepareDB(species = "Homo_sapiens", db = "MP", convert_species = TRUE, db_update = TRUE)
#> Species: Homo_sapiens
#> Warning: Use the mouse annotation to create the MP database for Homo_sapiens
#> Preparing database: MP
#> Convert species for the database: MP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset mmusculus_gene_ensembl ...
#> Match and convert the geneID...
#> 12939 genes mapped with ensembl_symbol
#> 10 genes mapped with entrez_symbol
#> 11 genes mapped with uniprot_symbol
#> ==============================
#> 12960 genes mapped
#> 71 genes unmapped
#> ==============================
#> Convert ID types for the database: MP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Match and convert the geneID...
#> 12843 genes mapped with ensembl_id
#> ==============================
#> 12843 genes mapped
#> 0 genes unmapped
#> ==============================
ListDB(species = "Homo_sapiens", db = "MP")
#>                                                         identifier version
#> 3 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                                                     comment  timestamp                date
#> 3 2023-01-16 08:00(converted from Mus_musculus) nterm:10107|Homo_sapiens-MP 1674032075 2023-01-18 08:54:34
#>                                                  db_version         db_name
#> 3 2023-01-16 08:00(converted from Mus_musculus) nterm:10107 Homo_sapiens-MP
#>                                                                    file
#> 3 /home/runner/.cache/R/R.cache/2c6309238a831951e934321463fcb730.Rcache
head(db_list[["Homo_sapiens"]][["MP"]][["TERM2GENE"]])
#>         Term      ensembl_id symbol entrez_id
#> 1 MP:0000600 ENSG00000139687    RB1      5925
#> 2 MP:0001716 ENSG00000139687    RB1      5925
#> 3 MP:0001698 ENSG00000139687    RB1      5925
#> 4 MP:0001092 ENSG00000139687    RB1      5925
#> 5 MP:0000961 ENSG00000139687    RB1      5925
#> 6 MP:0000245 ENSG00000139687    RB1      5925

db_list <- PrepareDB(species = "Macaca_fascicularis", db = "GO_BP", convert_species = TRUE, db_update = TRUE)
#> Species: Macaca_fascicularis
#> Install package: "org.Mf.eg.db" ...
#> 'getOption("repos")' replaces Bioconductor standard repositories, see '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> Bioconductor version 3.16 (BiocManager 1.30.19), R 4.2.2 (2022-10-31)
#> Installing package(s) 'org.Mf.eg.db'
#> Warning: package ‘org.Mf.eg.db’ is not available for Bioconductor version '3.16'
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Warning: package ‘org.Mf.eg.db’ is not available for this version of R
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Warning: Annotation package org.Mf.eg.db does not exist.
#> Warning: Use the human annotation to create the GO_BP database for Macaca_fascicularis
#> Preparing database: GO_BP
#> Convert species for the database: GO_BP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [feb2021.archive.ensembl.org:443] Operation timed out after 10000 milliseconds with 0 bytes received
ListDB(species = "Macaca_fascicularis", db = "GO_BP")
#> [1] identifier version    comment    timestamp  date       db_version db_name    file      
#> <0 rows> (or 0-length row.names)
head(db_list[["Macaca_fascicularis"]][["GO_BP"]][["TERM2GENE"]])
#> NULL

ccgenes <- CC_GenePrefetch("Homo_sapiens")
custom_TERM2GENE <- rbind(
  data.frame(term = "S_genes", gene = ccgenes[["cc_S_genes"]]),
  data.frame(term = "G2M_genes", gene = ccgenes[["cc_G2M_genes"]])
)
str(custom_TERM2GENE)
#> 'data.frame':	97 obs. of  2 variables:
#>  $ term: chr  "S_genes" "S_genes" "S_genes" "S_genes" ...
#>  $ gene: chr  "MCM5" "PCNA" "TYMS" "FEN1" ...
db_list <- PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"), db = "CellCycle", convert_species = TRUE,
  custom_TERM2GENE = custom_TERM2GENE, custom_species = "Homo_sapiens", custom_IDtype = "symbol", custom_version = "Seurat_v4"
)
#> Species: Homo_sapiens
#> Convert ID types for the database: CellCycle
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Error in curl::curl_fetch_memory(url, handle = handle): Timeout was reached: [feb2021.archive.ensembl.org:443] Operation timed out after 10001 milliseconds with 0 bytes received
ListDB(db = "CellCycle")
#>                                                         identifier version                                  comment
#> 7 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7 Seurat_v4 nterm:2|Homo_sapiens-CellCycle
#>    timestamp                date        db_version                db_name
#> 7 1674032268 2023-01-18 08:57:48 Seurat_v4 nterm:2 Homo_sapiens-CellCycle
#>                                                                    file
#> 7 /home/runner/.cache/R/R.cache/7eeb18e69d06066547bf0c0038774651.Rcache

db_list <- PrepareDB(species = "Mus_musculus", db = "CellCycle")
#> Species: Mus_musculus
head(db_list[["Mus_musculus"]][["CellCycle"]][["TERM2GENE"]])
#> NULL