Prepare the database for enrichment analysis

Prepare the database for enrichment analysis

Usage

PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"),
  db = c("GO_BP", "GO_CC", "GO_MF", "KEGG", "WikiPathway", "Reactome", "ProteinComplex",
    "DGI", "MP", "DO", "PFAM", "Chromosome", "GeneType", "Enzyme", "TF", "SP",
    "CellTalk", "CellChat"),
  db_IDtypes = c("symbol", "entrez_id", "ensembl_id"),
  db_version = "latest",
  db_update = FALSE,
  convert_species = FALSE,
  Ensembl_version = 103,
  mirror = NULL,
  custom_TERM2GENE = NULL,
  custom_TERM2NAME = NULL,
  custom_species = NULL,
  custom_IDtype = NULL,
  custom_version = NULL
)

Arguments

species

species

db

Enrichment database name.

db_update

Whether update the database to the latest version.

custom_version

Value

A list containing the database.

Examples

db_list <- PrepareDB(species = "Homo_sapiens", db = "GO_BP", db_update = TRUE)
#> Species: Homo_sapiens
#> Preparing database: GO_BP
#> Convert ID types for the database: GO_BP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Match and convert the geneID...
#> 18145 genes mapped with entrez_id
#> ==============================
#> 18145 genes mapped
#> 655 genes unmapped
#> ==============================
ListDB(species = "Homo_sapiens", db = "GO_BP")
#>                                                         identifier version                               comment
#> 3 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7 3.15.0 nterm:15947|Homo_sapiens-GO_BP
#>    timestamp                date         db_version            db_name
#> 3 1672928878 2023-01-05 14:27:57 3.15.0 nterm:15947 Homo_sapiens-GO_BP
#>                                                                    file
#> 3 /home/runner/.cache/R/R.cache/6479856b660afcefc0bbfe2b06217f1a.Rcache
head(db_list[["Homo_sapiens"]][["GO_BP"]][["TERM2GENE"]])
#>         Term entrez_id  symbol      ensembl_id
#> 1 GO:0000002      3980    LIG3 ENSG00000005156
#> 2 GO:0000002      9093  DNAJA3 ENSG00000103423
#> 3 GO:0000002      9093  DNAJA3 ENSG00000276726
#> 4 GO:0000002       291 SLC25A4 ENSG00000151729
#> 5 GO:0000002     56652    TWNK ENSG00000107815
#> 6 GO:0000002     30968  STOML2 ENSG00000165283

db_list <- PrepareDB(species = "Homo_sapiens", db = "MP", convert_species = TRUE, db_update = TRUE)
#> Species: Homo_sapiens
#> Warning: Use the mouse annotation to create the MP database for Homo_sapiens
#> Preparing database: MP
#> Convert species for the database: MP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset mmusculus_gene_ensembl ...
#> Match and convert the geneID...
#> 12859 genes mapped with ensembl_symbol
#> 10 genes mapped with entrez_symbol
#> 9 genes mapped with uniprot_symbol
#> ==============================
#> 12878 genes mapped
#> 62 genes unmapped
#> ==============================
#> Convert ID types for the database: MP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Match and convert the geneID...
#> 12764 genes mapped with ensembl_id
#> ==============================
#> 12764 genes mapped
#> 0 genes unmapped
#> ==============================
ListDB(species = "Homo_sapiens", db = "MP")
#>                                                          identifier version
#> 12 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                                                      comment  timestamp                date
#> 12 2023-01-02 08:00(converted from Mus_musculus) nterm:10108|Homo_sapiens-MP 1672928967 2023-01-05 14:29:27
#>                                                   db_version         db_name
#> 12 2023-01-02 08:00(converted from Mus_musculus) nterm:10108 Homo_sapiens-MP
#>                                                                     file
#> 12 /home/runner/.cache/R/R.cache/e1ca3baa980aea6fb491fe760a492dd2.Rcache
head(db_list[["Homo_sapiens"]][["MP"]][["TERM2GENE"]])
#>         Term      ensembl_id symbol entrez_id
#> 1 MP:0000600 ENSG00000139687    RB1      5925
#> 2 MP:0001716 ENSG00000139687    RB1      5925
#> 3 MP:0001698 ENSG00000139687    RB1      5925
#> 4 MP:0001092 ENSG00000139687    RB1      5925
#> 5 MP:0000961 ENSG00000139687    RB1      5925
#> 6 MP:0000245 ENSG00000139687    RB1      5925

db_list <- PrepareDB(species = "Macaca_fascicularis", db = "GO_BP", convert_species = TRUE, db_update = TRUE)
#> Species: Macaca_fascicularis
#> Install package: "org.Mf.eg.db" ...
#> 'getOption("repos")' replaces Bioconductor standard repositories, see '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> Bioconductor version 3.16 (BiocManager 1.30.19), R 4.2.2 (2022-10-31)
#> Installing package(s) 'org.Mf.eg.db'
#> Warning: package ‘org.Mf.eg.db’ is not available for Bioconductor version '3.16'
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Warning: package ‘org.Mf.eg.db’ is not available for this version of R
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Warning: Annotation package org.Mf.eg.db does not exist.
#> Warning: Use the human annotation to create the GO_BP database for Macaca_fascicularis
#> Preparing database: GO_BP
#> Convert species for the database: GO_BP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Ensembl site unresponsive, trying useast mirror
#> Match and convert the geneID...
#> 18145 genes mapped with entrez_id
#> ==============================
#> 18145 genes mapped
#> 655 genes unmapped
#> ==============================
#> Convert ID types for the database: GO_BP
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset mfascicularis_gene_ensembl ...
#> Ensembl site unresponsive, trying useast mirror
#> Match and convert the geneID...
#> 15803 genes mapped with ensembl_id
#> ==============================
#> 15803 genes mapped
#> 0 genes unmapped
#> ==============================
ListDB(species = "Macaca_fascicularis", db = "GO_BP")
#>                                                          identifier version
#> 13 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                                                      comment  timestamp                date
#> 13 3.15.0(converted from Homo_sapiens) nterm:15713|Macaca_fascicularis-GO_BP 1672929321 2023-01-05 14:35:21
#>                                         db_version                   db_name
#> 13 3.15.0(converted from Homo_sapiens) nterm:15713 Macaca_fascicularis-GO_BP
#>                                                                     file
#> 13 /home/runner/.cache/R/R.cache/f13510a82f20793bcec9de7f15b49a8c.Rcache
head(db_list[["Macaca_fascicularis"]][["GO_BP"]][["TERM2GENE"]])
#>         Term         ensembl_id   symbol entrez_id
#> 1 GO:0000002 ENSMFAG00000036202     LIG3 102140711
#> 2 GO:0000002 ENSMFAG00000042573   DNAJA3 102143112
#> 3 GO:0000002 ENSMFAG00000032688  SLC25A4 102120762
#> 4 GO:0000002 ENSMFAG00000032824     TWNK 102131591
#> 5 GO:0000002 ENSMFAG00000039141   STOML2 102124083
#> 6 GO:0000002 ENSMFAG00000000534 PPARGC1A 102125865

ccgenes <- CC_GenePrefetch("Homo_sapiens")
custom_TERM2GENE <- rbind(
  data.frame(term = "S_genes", gene = ccgenes[["cc_S_genes"]]),
  data.frame(term = "G2M_genes", gene = ccgenes[["cc_G2M_genes"]])
)
str(custom_TERM2GENE)
#> 'data.frame':	97 obs. of  2 variables:
#>  $ term: chr  "S_genes" "S_genes" "S_genes" "S_genes" ...
#>  $ gene: chr  "MCM5" "PCNA" "TYMS" "FEN1" ...
db_list <- PrepareDB(
  species = c("Homo_sapiens", "Mus_musculus"), db = "CellCycle", convert_species = TRUE,
  custom_TERM2GENE = custom_TERM2GENE, custom_species = "Homo_sapiens", custom_IDtype = "symbol", custom_version = "Seurat_v4"
)
#> Species: Homo_sapiens
#> Convert ID types for the database: CellCycle
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Ensembl site unresponsive, trying useast mirror
#> Match and convert the geneID...
#> 97 genes mapped with ensembl_symbol
#> ==============================
#> 97 genes mapped
#> 0 genes unmapped
#> ==============================
#> Species: Mus_musculus
#> Warning: Use the Homo_sapiens annotation to create the CellCycle database for Mus_musculus
#> Convert species for the database: CellCycle
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset hsapiens_gene_ensembl ...
#> Match and convert the geneID...
#> 97 genes mapped with ensembl_symbol
#> ==============================
#> 97 genes mapped
#> 0 genes unmapped
#> ==============================
#> Convert ID types for the database: CellCycle
#> Connect to the Ensembl archives...
#> Using the 103 version of biomart...
#> Connect to the biomart...
#> Connect to the dataset mmusculus_gene_ensembl ...
#> Match and convert the geneID...
#> 94 genes mapped with ensembl_id
#> ==============================
#> 94 genes mapped
#> 0 genes unmapped
#> ==============================
ListDB(db = "CellCycle")
#>                                                         identifier version
#> 6 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 1 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                                                 comment  timestamp                date
#> 6                              Seurat_v4 nterm:2|Homo_sapiens-CellCycle 1672929348 2023-01-05 14:35:48
#> 1 Seurat_v4(converted from Homo_sapiens) nterm:2|Mus_musculus-CellCycle 1672929371 2023-01-05 14:36:10
#>                                       db_version                db_name
#> 6                              Seurat_v4 nterm:2 Homo_sapiens-CellCycle
#> 1 Seurat_v4(converted from Homo_sapiens) nterm:2 Mus_musculus-CellCycle
#>                                                                    file
#> 6 /home/runner/.cache/R/R.cache/7eeb18e69d06066547bf0c0038774651.Rcache
#> 1 /home/runner/.cache/R/R.cache/0a6a5861c4521ca291de6c96159106f5.Rcache

db_list <- PrepareDB(species = "Mus_musculus", db = "CellCycle")
#> Species: Mus_musculus
#> Loaded cached db: CellCycle version:Seurat_v4(converted from Homo_sapiens) nterm:2 created:2023-01-05 14:36:10
head(db_list[["Mus_musculus"]][["CellCycle"]][["TERM2GENE"]])
#>      Term         ensembl_id symbol entrez_id
#> 1 S_genes ENSMUSG00000005410   Mcm5     17218
#> 2 S_genes ENSMUSG00000027342   Pcna     18538
#> 3 S_genes ENSMUSG00000025747   Tyms     22171
#> 4 S_genes ENSMUSG00000024742   Fen1     14156
#> 5 S_genes ENSMUSG00000029730   Mcm7     17220
#> 6 S_genes ENSMUSG00000022673   Mcm4     17217