Run cell-level quality control for single cell RNA-seq data.

In CellQC, doublet-calling will be run first and then reject abnormal cell data based on median absolute deviation (MAD) outliers. After doublet-calling and outlier filtering, CellQC will perform general QC (gene count, UMI count, etc.) and reject cell data for non-species of interest based on the proportion and number of UMIs for the species.

Usage

RunCellQC(
  srt,
  assay = "RNA",
  qc_metrics = c("doublets", "outlier", "umi", "gene", "mito", "ribo", "ribo_mito_ratio",
    "species"),
  return_filtered = FALSE,
  db_method = "scDblFinder",
  db_rate = ncol(srt)/1000 * 0.01,
  outlier_cutoff = c("log10_nCount:lower:2.5", "log10_nCount:higher:5",
    "log10_nFeature:lower:2.5", "log10_nFeature:higher:5", "featurecount_dist:lower:2.5"),
  outlier_n = 1,
  UMI_threshold = 3000,
  gene_threshold = 1000,
  mito_threshold = 20,
  mito_pattern = c("MT-", "Mt-", "mt-"),
  mito_gene = NULL,
  ribo_threshold = 50,
  ribo_pattern = c("RP[SL]\\d+\\w{0,1}\\d*$", "Rp[sl]\\d+\\w{0,1}\\d*$",
    "rp[sl]\\d+\\w{0,1}\\d*$"),
  ribo_gene = NULL,
  ribo_mito_ratio_range = c(1, Inf),
  species = NULL,
  species_gene_prefix = NULL,
  species_percent = 95,
  seed = 11
)

Arguments

srt

A Seurat object.

assay

Assay to use.

qc_metrics

QC metrics applied.

return_filtered

Whether to return a cell-filtered Seurat object.

db_method

Doublet-calling methods used. Can be one of scDblFinder, Scrublet, DoubletDetection, scds_cxds, scds_bcds, scds_hybrid

db_rate

The expected doublet rate. By default this is assumed to be 1% per thousand cells captured (so 4% among 4000 thousand cells), which is appropriate for 10x datasets.

outlier_cutoff

MAD outlier metrics. See isOutlier.

outlier_n

Minimum number of outlier metrics that meet the conditions for determining outlier cells. Default is 1.

UMI_threshold

UMI number threshold. Cells that exceed this threshold will be considered as kept. Default is 3000.

gene_threshold

Gene number threshold. Cells that exceed this threshold will be considered as kept. Default is 1000.

mito_threshold

Percentage of UMI counts of mitochondrial genes. Cells that exceed this threshold will be considered as discarded. Default is 20.

mito_pattern

Regex patterns to match the mitochondrial genes.

mito_gene

A defined mitochondrial genes. If features provided, will ignore the mito_pattern matching.

ribo_threshold

Percentage of UMI counts of ribosomal genes. Cells that exceed this threshold will be considered as discarded. Default is 50.

ribo_pattern

Regex patterns to match the ribosomal genes.

ribo_gene

A defined ribosomal genes. If features provided, will ignore the ribo_pattern matching.

species

Species used as the suffix of the QC metrics. The first is the species of interest. Default is NULL.

species_gene_prefix

Species gene prefix used to calculate QC metrics for each species. Default is NULL.

species_percent

Percentage of UMI counts of the first species. Cells that exceed this threshold will be considered as kept. Default is 95.

seed

Set a random seed. Default is 11.

Value

Returns Seurat object with the QC results stored in the meta.data slot.

Note

General quality control usually uses metrics such as gene count, UMI count, etc. In addition, ribo content and mito content can be used as QC indicators: mito content can be used as an indication of apoptosis, with a general threshold of 20 ribo content is cell type dependent, with some cell types having even more than 50 however, ribo content in single cell data can also indicate whether the empty droplets, and the ribo content in empty droplets is usually high and the mito content is low instead.

Examples

data("pancreas_sub")
pancreas_sub <- RunCellQC(pancreas_sub)
#> Run doublet-calling with scDblFinder
#> Calculate outlier cells
#>      log10_nCount:lower:2.5       log10_nCount:higher:5    log10_nFeature:lower:2.5     log10_nFeature:higher:5 
#>                           0                           0                           5                           0 
#> featurecount_dist:lower:2.5 
#>                           8 
#> >>> Total cells: 1000 
#> >>> Cells which are filtered out: 35 
#> ... 23 potential doublets 
#> ... 12 outlier cells 
#> ... 0 low-UMI cells 
#> ... 0 low-gene cells 
#> ... 0 high-mito cells 
#> ... 0 high-ribo cells 
#> ... 0  ribo_mito_ratio outlier cells 
#> ... 0 species-contaminated cells 
#> >>> Remained cells after filtering: 965