Run Non-negative matrix factorization(NMF)

Run a NMF dimensionality reduction.

Usage

RunNMF(object, ...)

# S3 method for Seurat
RunNMF(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "nmf",
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 42,
  ...
)

# S3 method for Assay
RunNMF(
  object,
  assay = NULL,
  slot = "data",
  features = NULL,
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 42,
  ...
)

# S3 method for default
RunNMF(
  object,
  assay = NULL,
  slot = "data",
  nbes = 50,
  nmf.method = "RcppML",
  tol = 1e-05,
  maxit = 100,
  rev.nmf = FALSE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.key = "BE_",
  verbose = TRUE,
  seed.use = 42,
  ...
)

Arguments

object

An object

...

assay

Name of Assay nmf is being run on.

slot

Name of slot nmf is being run on.

features

Features to compute nmf on. If features=NULL, nmf will be run using the variable features for the Assay.

nbes

Total Number of BEs("basis experiment", aka "metagene") to compute and store (50 by default).

rev.nmf

By default computes the nmf on the cell x gene matrix. Setting to true will compute it on gene x cell matrix.

ndims.print

BEs to print genes for.

nfeatures.print

Number of genes to print for each BE.

reduction.name

dimensional reduction name, 'nmf' by default

reduction.key

dimensional reduction key, specifies the string before the number for the dimension names. "BE_" by default.

verbose

Print the top genes associated with high/low loadings for the BEs.

seed.use

Set a random seed. By default, sets the seed to 42. Setting NULL will not set a seed.

Value

Returns Seurat object with the NMF calculation stored in the reductions slot.