SCP palette

The default palette for SCP plot function.

Usage

palette_scp(
  x,
  n = 100,
  palette = "Paired",
  palcolor = NULL,
  type = "auto",
  matched = FALSE,
  reverse = FALSE,
  NA_keep = FALSE,
  NA_color = "grey80"
)

Arguments

x

A vector of character/factor or numeric values for color mapping.

palette

palette name used.

palcolor

type

Type of x.Can be one of "auto","discrete" or "continuous". The default is "auto", which automatically detects if x is a numeric value. If it is numeric, then x is interpolated to the palette color of length 100 in ascending order. Otherwise, the color is assigned to x according to its type number.

matched

Whether to return a matched color vector of length x

reverse

Whether to invert the colors.

NA_keep

Whether to keep the color assignment to NA in x.

NA_color

NA colors if NA_keep is TRUE.

Examples

x <- c(1:3, NA, 3:5)
palette_scp(x, palette = "Spectral")
#>    [1,1.04] (1.04,1.08] (1.08,1.12] (1.12,1.16]  (1.16,1.2]  (1.2,1.24] (1.24,1.28] (1.28,1.32] (1.32,1.36]  (1.36,1.4] 
#>   "#5E4FA2"   "#5954A4"   "#555AA7"   "#5060AA"   "#4C66AC"   "#476BAF"   "#4371B2"   "#3E77B5"   "#3A7DB7"   "#3682BA" 
#>  (1.4,1.44] (1.44,1.48] (1.48,1.52] (1.52,1.56]  (1.56,1.6]  (1.6,1.64] (1.64,1.68] (1.68,1.72] (1.72,1.76]  (1.76,1.8] 
#>   "#3288BC"   "#378EBA"   "#3D94B7"   "#429AB5"   "#47A0B3"   "#4CA5B0"   "#52ABAE"   "#57B1AB"   "#5CB7A9"   "#61BDA6" 
#>  (1.8,1.84] (1.84,1.88] (1.88,1.92] (1.92,1.96]    (1.96,2]    (2,2.04] (2.04,2.08] (2.08,2.12] (2.12,2.16]  (2.16,2.2] 
#>   "#67C2A4"   "#6EC5A4"   "#75C8A4"   "#7CCAA4"   "#83CDA4"   "#8AD0A4"   "#91D2A4"   "#98D5A4"   "#9FD8A4"   "#A6DBA4" 
#>  (2.2,2.24] (2.24,2.28] (2.28,2.32] (2.32,2.36]  (2.36,2.4]  (2.4,2.44] (2.44,2.48] (2.48,2.52] (2.52,2.56]  (2.56,2.6] 
#>   "#ACDDA3"   "#B2E0A2"   "#B8E2A1"   "#BEE5A0"   "#C4E79E"   "#CAE99D"   "#D0EC9C"   "#D6EE9B"   "#DCF199"   "#E2F398" 
#>  (2.6,2.64] (2.64,2.68] (2.68,2.72] (2.72,2.76]  (2.76,2.8]  (2.8,2.84] (2.84,2.88] (2.88,2.92] (2.92,2.96]    (2.96,3] 
#>   "#E7F599"   "#E9F69D"   "#ECF7A1"   "#EEF8A5"   "#F1F9A9"   "#F3FAAD"   "#F6FBB1"   "#F8FCB5"   "#FBFDB9"   "#FDFEBD" 
#>    (3,3.04] (3.04,3.08] (3.08,3.12] (3.12,3.16]  (3.16,3.2]  (3.2,3.24] (3.24,3.28] (3.28,3.32] (3.32,3.36]  (3.36,3.4] 
#>   "#FEFDBC"   "#FEFAB7"   "#FEF7B1"   "#FEF4AC"   "#FEF0A7"   "#FEEDA2"   "#FEEA9C"   "#FEE797"   "#FEE492"   "#FEE18D" 
#>  (3.4,3.44] (3.44,3.48] (3.48,3.52] (3.52,3.56]  (3.56,3.6]  (3.6,3.64] (3.64,3.68] (3.68,3.72] (3.72,3.76]  (3.76,3.8] 
#>   "#FDDC88"   "#FDD784"   "#FDD27F"   "#FDCD7B"   "#FDC877"   "#FDC373"   "#FDBE6F"   "#FDB96A"   "#FDB466"   "#FDAF62" 
#>  (3.8,3.84] (3.84,3.88] (3.88,3.92] (3.92,3.96]    (3.96,4]    (4,4.04] (4.04,4.08] (4.08,4.12] (4.12,4.16]  (4.16,4.2] 
#>   "#FCA95E"   "#FBA25B"   "#FA9C58"   "#F99555"   "#F88F52"   "#F7884F"   "#F6824C"   "#F67B49"   "#F57446"   "#F46E43" 
#>  (4.2,4.24] (4.24,4.28] (4.28,4.32] (4.32,4.36]  (4.36,4.4]  (4.4,4.44] (4.44,4.48] (4.48,4.52] (4.52,4.56]  (4.56,4.6] 
#>   "#F16943"   "#EE6445"   "#EB5F46"   "#E85A47"   "#E45648"   "#E1514A"   "#DE4C4B"   "#DB474C"   "#D8434D"   "#D53E4E" 
#>  (4.6,4.64] (4.64,4.68] (4.68,4.72] (4.72,4.76]  (4.76,4.8]  (4.8,4.84] (4.84,4.88] (4.88,4.92] (4.92,4.96]    (4.96,5] 
#>   "#CF384D"   "#CA324C"   "#C42C4B"   "#BF2549"   "#B91F48"   "#B41947"   "#AE1345"   "#A90D44"   "#A30743"   "#9E0142" 
palette_scp(x, palette = "Spectral", matched = TRUE)
#>  [1,1.04]  (1.96,2]  (2.96,3]      <NA>  (2.96,3]  (3.96,4]  (4.96,5] 
#> "#5E4FA2" "#83CDA4" "#FDFEBD"        NA "#FDFEBD" "#F88F52" "#9E0142" 
palette_scp(x, palette = "Spectral", matched = TRUE, NA_keep = TRUE)
#>  [1,1.04]  (1.96,2]  (2.96,3]      <NA>  (2.96,3]  (3.96,4]  (4.96,5] 
#> "#5E4FA2" "#83CDA4" "#FDFEBD"  "grey80" "#FDFEBD" "#F88F52" "#9E0142" 
palette_scp(x, palette = "Paired", type = "discrete")
#>         1         2         3         4         5 
#> "#A6CEE3" "#1F78B4" "#B2DF8A" "#33A02C" "#FDBF6F" 

if (FALSE) {
if (interactive()) {
  check_R(c("stringr", "RColorBrewer", "ggsci", "Redmonder", "rcartocolor", "nord", "viridis", "pals", "oompaBase"))
  library(stringr)
  library(RColorBrewer)
  library(ggsci)
  library(Redmonder)
  library(rcartocolor)
  library(nord)
  library(viridis)
  library(pals)
  brewer.pal.info <- RColorBrewer::brewer.pal.info
  ggsci_db <- ggsci:::ggsci_db
  redmonder.pal.info <- Redmonder::redmonder.pal.info
  metacartocolors <- rcartocolor::metacartocolors
  rownames(metacartocolors) <- metacartocolors$Name
  nord_palettes <- nord::nord_palettes
  viridis_names <- c("magma", "inferno", "plasma", "viridis", "cividis", "rocket", "mako", "turbo")
  viridis_palettes <- lapply(setNames(viridis_names, viridis_names), function(x) viridis::viridis(100, option = x))
  ocean_names <- names(pals:::syspals)[str_detect(names(pals:::syspals), "ocean")]
  ocean_palettes <- pals:::syspals[ocean_names]
  custom_names <- c("jet", "simspec")
  custom_palettes <- list(
    oompaBase::jetColors(N = 100),
    c("#c22b86", "#f769a1", "#fcc5c1", "#253777", "#1d92c0", "#9ec9e1", "#015b33", "#42aa5e", "#d9f0a2", "#E66F00", "#f18c28", "#FFBB61")
  )
  names(custom_palettes) <- custom_names

  palette_list <- list()
  all_colors <- c(
    rownames(brewer.pal.info), names(ggsci_db), rownames(redmonder.pal.info),
    rownames(metacartocolors), names(nord_palettes), names(viridis_palettes),
    ocean_names, custom_names
  )
  for (pal in all_colors) {
    if (!pal %in% all_colors) {
      stop(paste0("Invalid pal Must be one of ", paste0(all_colors, collapse = ",")))
    }
    if (pal %in% rownames(brewer.pal.info)) {
      pal_n <- brewer.pal.info[pal, "maxcolors"]
      pal_category <- brewer.pal.info[pal, "category"]
      if (pal_category == "div") {
        palcolor <- rev(brewer.pal(name = pal, n = pal_n))
      } else {
        if (pal == "Paired") {
          palcolor <- brewer.pal(12, "Paired")[c(1:4, 7, 8, 5, 6, 9, 10, 11, 12)]
        } else {
          palcolor <- brewer.pal(name = pal, n = pal_n)
        }
      }
    } else if (pal %in% names(ggsci_db)) {
      if (pal %in% c("d3", "uchicago", "material")) {
        for (subpal in names(ggsci_db[[pal]])) {
          palcolor <- ggsci_db[[pal]][[subpal]]
          palette_list[[paste0(pal, "-", subpal)]] <- palcolor
        }
        next
      } else {
        palcolor <- ggsci_db[[pal]][[1]]
      }
    } else if (pal %in% rownames(redmonder.pal.info)) {
      pal_n <- redmonder.pal.info[pal, "maxcolors"]
      pal_category <- redmonder.pal.info[pal, "category"]
      if (pal_category == "div") {
        palcolor <- rev(redmonder.pal(name = pal, n = pal_n))
      } else {
        palcolor <- redmonder.pal(name = pal, n = pal_n)
      }
    } else if (pal %in% rownames(metacartocolors)) {
      pal_n <- metacartocolors[pal, "Max_n"]
      palcolor <- carto_pal(name = pal, n = pal_n)
    } else if (pal %in% names(nord_palettes)) {
      palcolor <- nord_palettes[[pal]]
    } else if (pal %in% names(viridis_palettes)) {
      palcolor <- viridis_palettes[[pal]]
    } else if (pal %in% names(ocean_palettes)) {
      palcolor <- ocean_palettes[[pal]]
    } else if (pal %in% custom_names) {
      palcolor <- custom_palettes[[pal]]
    }
    palette_list[[pal]] <- palcolor
  }
  # save(palette_list, file = "~/Git/SCP/R/sysdata.rda", version = 2)
}
}