Read single-cell proteomics data as a QFeatures object from tabular data and metadata

Convert tabular quantitative MS data and metadata from a spreadsheet or a data.frame into a QFeatures object containing SingleCellExperiment objects.

readSCP(featureData, colData, batchCol, channelCol, verbose = TRUE, ...)

Arguments

featureData
File or object holding the quantitative data. Can be either a character(1) with the path to a text-based spreadsheet (comma-separated values by default, but see ...) or an object that can be coerced to a data.frame. It is advised not to encode characters as factors.

colData
A data.frame or any object that can be coerced to a data.frame. colData is expected to contains all the sample meta information. Required fields are the acquisition batch (given by batchCol) and the acquisition channel within the batch (e.g. TMT channel, given by channelCol). Additional fields (e.g. sample type, acquisition date,...) are allowed and will be stored as sample meta data.

batchCol
A numeric(1) or character(1) pointing to the column of featureData and colData that contain the batch names. Make sure that the column name in both table are either identical (if you supply a character) or have the same index (if you supply a numeric).

channelCol
A numeric(1) or character(1) pointing to the column of colData that contains the column names of the quantitive data in featureData (see Example).

verbose
A logical(1) indicating whether the progress of the data reading and formatting should be printed to the console. Default is TRUE.

...
Further arguments that can be passed on to read.csv except stringsAsFactors, which is always FALSE.

Value

An instance of class QFeatures. The expression data of each batch is stored in a separate assay as a SingleCellExperiment object.

Note

The SingleCellExperiment class is built on top of the RangedSummarizedExperiment class. This means that some column names are forbidden in the rowData. Avoid using the following names: seqnames, ranges, strand, start, end, width, element

Author

Laurent Gatto, Christophe Vanderaa

Examples

## Load an example table containing MaxQuant output
data("mqScpData")

## Load the (user-generated) annotation table
data("sampleAnnotation")

## Format the tables into a QFeatures object
readSCP(featureData = mqScpData,
        colData = sampleAnnotation,
        batchCol = "Set",
        channelCol = "Channel")

#> Loading data as a 'SingleCellExperiment' object

#> Splitting data based on 'Set'

#> Formatting sample metadata (colData)

#> Formatting data as a 'QFeatures' object

#> An instance of class QFeatures containing 4 assays:
#>  [1] 190222S_LCA9_X_FP94BM: SingleCellExperiment with 334 rows and 16 columns 
#>  [2] 190321S_LCA10_X_FP97_blank_01: SingleCellExperiment with 109 rows and 16 columns 
#>  [3] 190321S_LCA10_X_FP97AG: SingleCellExperiment with 433 rows and 16 columns 
#>  [4] 190914S_LCB3_X_16plex_Set_21: SingleCellExperiment with 321 rows and 16 columns